### 2025-09-29 #9096 @editorialbot
[PRE REVIEW]: `peaks`: a Python package for analysis of angle-resolved photoemission and related spectroscopies
**Submitting author:** @phrgab (<a href="http://orcid.org/0000-0002-6523-9034">P D C King</a>)
**Repository:** https://github.com/phrgab/peaks
**Branch with paper.md** (empty if default branch):
**Version:** v0.4.0
