### 2025-09-30 #9096 @phrgab
Thanks @kyleniemeyer

For reviewer suggestions, from the list of existing JOSS reviewers I would suggest considering:
- maurov
- stuartcampbell
- jsdodge
- tillsten
- ivergara

Also could consider kmnhan who is not on your list I don't think, but is more (very) closely aligned in terms of domain area.
