### 2025-10-14 cf54c3a
fix unit handling in merging of photon energy scans
