+++ pyproject.toml
@@ -16,7 +16,7 @@ classifiers = [
     "Topic :: Scientific/Engineering :: Physics",
 ]
 keywords = ["ARPES", "photoemission", "spectroscopy", "xarray"]
-requires-python = ">=3.11"
+requires-python = ">=3.11,<3.14"
 dependencies = [
         "ipykernel (>=6.25,<7)",
         "ipywidgets (>=8.1,<9)",
