### 2026-01-16 4f1157a
bump version to 0.4.6.dev and update changelog
