+++ CHANGELOG.md
@@ -4,6 +4,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.4.5]
+
+### Fixed
+- Failures when mergeing two FSMs
+- Pinned numba version to >=0.60.0 due to dependency requirements
+
 ## [0.4.4]
 
 ### Fixed
