### 2026-01-16 86e997e
previou commit: register units after interpolation in summing function
