+++ peaks/core/utils/sample_data.py
@@ -41,7 +41,7 @@ class ZenodoDownloader:
         retries = Retry(
             total=5,
             backoff_factor=1,
-            status_forcelist=[502, 503, 504],
+            status_forcelist=[429, 502, 503, 504],
             allowed_methods=["GET"],
             raise_on_status=False,
         )
