+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Fixed
 - Bug when summing two FSMs
 - Pinned numba version to >=0.60.0 to resolve dependency conflicts
+- Downloading errors due to rate limiting
 
 ## [0.4.4]
 
