### 2026-01-19 a71285a
bump version properly to 0.4.5
