### 2026-01-22 46542c6
add custom User-Agent and small delay for zenodo downloads
