### 2026-01-22 1c462c9
bump version to 0.4.6, update changelog and fix lmfit intersphix url
