+++ peaks/__init__.py
@@ -17,7 +17,7 @@ To load data, use the load function::
 
 """
 
-__version__ = "0.4.5"
+__version__ = "0.4.6"
 
 # Set some default xarray options
 import xarray as xr
