+++ CHANGELOG.md
@@ -4,6 +4,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.4.6]
+
+### Fixed
+- Zenodo download failures
+- Excluded yanked `uncertainties` version 3.2.4
+- `lmfit` intersphinx URL for doc building
+
 ## [0.4.5]
 
 ### Fixed
