### 2026-01-23 73ccaa8
bump main version to 0.4.7.dev
