### 2026-01-23 589a396
average out floating point error in GUI step size calc
