### 2026-01-26 c6ad5be
bump version to 0.4.7 and update changelog
