### 2026-01-27 6745a86
Auto trim non-monotonic data in ana-polar maps
