### 2026-02-05 5a185ad
Added Loader for SGM4
Added new data loader for the SGM4 beamline in Aarhus.
A new script SGM4peaksloader.py has been added to the loaders folder containing the SGM4 loader definition.
The corresponding init file has been updated so the SGM4loader is loaded with the rest of the library.
The loc_registry.py script has been updated so that SGM4 files are automatically identified as such, and loaded with the SGM4 loader, when passed to the generic peaks.load() function.
