+++ peaks/core/fileIO/loaders/SGM4peaksloader.py
@@ -1,0 +1,640 @@
+import math
+from typing import Optional, Union
+
+import h5py
+import numpy as np
+
+# import pint
+import pint_xarray
+import xarray as xr
+
+from peaks.core.fileIO.base_arpes_data_classes.base_arpes_data_class import (
+    BaseARPESDataLoader,
+)
+from peaks.core.fileIO.base_data_classes.base_hdf5_class import BaseHDF5DataLoader
+
+# from peaks.core.fileIO.base_data_classes.base_hdf5_class import BaseDataLoader
+from peaks.core.fileIO.loc_registry import register_loader
+from peaks.core.metadata.base_metadata_models import BaseMetadataModel, Quantity
+
+ureg = pint_xarray.unit_registry
+
+
+class SGM4NanoFocussingMetadataModel(BaseMetadataModel):
+    """Model to store metadata for SGM4 Nano-ARPES focussing optics metadata."""
+
+    CapX: Optional[Union[str, Quantity]] = None
+    CapY: Optional[Union[str, Quantity]] = None
+    CapZ: Optional[Union[str, Quantity]] = None
+
+
+class SGM4TemperatureMetadataModel(BaseMetadataModel):
+    """Model to store metadata for SGM4 Nano-ARPES focussing optics metadata."""
+
+    ColdHead: Optional[Union[str, Quantity]] = None
+    HeaterPower: Optional[Union[str, Quantity]] = None
+    SampleStage: Optional[Union[str, Quantity]] = None
+    Stinger: Optional[Union[str, Quantity]] = None
+
+
+@register_loader
+# class SGM4NanoARPESLoader(BaseHDF5DataLoader):
+class SGM4NanoARPESLoader(BaseHDF5DataLoader, BaseARPESDataLoader):  # ,
+    _loc_name = "SGM4"
+    _loc_description = (
+        "SGM4 endstation at the ASTRID2 synchrotron at Aarhus University, Denmark."
+    )
+    _loc_url = "https://isa.au.dk/facilities/astrid2/beamlines/AU-sgm4/AU-sgm4.asp"
+    _analyser_slit_angle = 0 * ureg.deg
+    _analyser_WF = 4.416 * ureg.electron_volt
+    _anapolaroffset = -14 * ureg.deg
+    _smpolaroffset = +45 * ureg.deg
+
+    _unitdict = {
+        "x1": ureg.micrometer,
+        "x2": ureg.micrometer,
+        "x3": ureg.micrometer,
+        "polar": ureg.deg,
+        "tilt": ureg.deg,
+        "azi": ureg.deg,
+        "defocus": ureg.micrometer,
+        "CapZ": ureg.micrometer,
+        "CapY": ureg.micrometer,
+        "CapX": ureg.micrometer,
+        "SamX": ureg.micrometer,
+        "SamY": ureg.micrometer,
+        "SamZ": ureg.micrometer,
+        "CapPolar": ureg.deg,
+        "eV": ureg.electron_volt,
+        "theta_par": ureg.deg,
+        "deflector_perp": ureg.deg,
+        "deflector_parallel": ureg.deg,
+        "Loop": ureg.count,
+    }
+
+    _manipulator_axes = ["polar", "tilt", "azi", "x1", "x2", "x3", "defocus"]
+    _manipulator_name_conventions = {
+        "polar": "SamPolar",
+        "tilt": "SamTilt",
+        "azi": "SamAzi",
+        "x1": "FSamX",
+        "x2": "FSamY",
+        "x3": "FSamZ",
+        "defocus": "CapZ",
+    }
+
+    _manipulator_sign_conventions = {
+        # HR-branch
+        "polar": +1,
+        "tilt": -1,
+    }
+
[…]
