+++ peaks/core/fileIO/loaders/__init__.py
@@ -2,6 +2,7 @@
 
 # ruff: noqa: F401
 from peaks.core.fileIO.loaders import (
+    SGM4peaksloader,
     ape,
     bloch,
     casiopee,
