### 2026-03-02 6e39a5a
remove slant correction in docs; add local path support for the cif data
