### 2026-03-02 2405c4d
remove numba version constraint as it's no longer causing conflicts
