### 2026-03-04 3386490
test installs regularly across various python versions and platforms using CI
