+++ .github/workflows/test-installs.yml
@@ -1,0 +1,32 @@
+name: Install test
+
+on:
+  schedule:
+    - cron: "30 8 * * 1" # every Monday at 0830 UTC
+  workflow_dispatch: # can be triggered manually from the GH UI
+
+jobs:
+  install-test:
+    runs-on: ${{ matrix.os }}
+
+    strategy:
+      fail-fast: false # don't cancel other jobs if one fails here
+      matrix:
+        os: [ubuntu-latest, macos-latest, windows-latest]
+        python-version: ["3.12", "3.13", "3.14"]
+
+    steps:
+      - name: Check out the repo
+        uses: actions/checkout@v6
+
+      - name: Set up python ${{ matrix.python-version }}
+        uses: actions/setup-python@v6
+        with:
+          python-version: ${{ matrix.python-version }}
+          cache: "pip"
+
+      - name: Install peaks
+        run: pip install .
+
+      - name: Test import
+        run: python -c "import peaks"
