### 2026-03-04 5cf2cba
use the latest version (v6) of GH actions (checkout and setup-python)
