### 2026-03-04 07c7033
add a push trigger to test the CI
