### 2026-03-04 50a7f92
add the structure module to install tests
