### 2026-03-04 e7dc264
include python 3.11 for install tests
