+++ .github/workflows/test-installs.yml
@@ -15,7 +15,7 @@ jobs:
       fail-fast: false # don't cancel other jobs if one fails here
       matrix:
         os: [ubuntu-latest, macos-latest, windows-latest]
-        python-version: ["3.12", "3.13", "3.14"]
+        python-version: ["3.11", "3.12", "3.13", "3.14"]
 
     steps:
       - name: Check out the repo
