### 2026-03-04 6253d1e
specify the repo so it only runs in the original peaks repo and add badge for it
