+++ README.md
@@ -2,14 +2,15 @@
 
 [![git](https://img.shields.io/badge/repo-github-blue?logo=github)](https://github.com/phrgab/peaks)
 [![Latest version on pypi](https://img.shields.io/pypi/v/peaks-arpes?logo=pypi&logoColor=white&color=brightgree)](https://pypi.org/project/peaks-arpes/)
+[![GitHub Actions Workflow Status](https://img.shields.io/github/actions/workflow/status/phrgab/peaks/test-installs.yml?logo=github-actions&logoColor=white&label=Install%20tests)](https://github.com/phrgab/peaks/actions/workflows/test-installs.yml)
 [![Supported Python versions](https://img.shields.io/pypi/pyversions/peaks-arpes?logo=python&logoColor=white)](https://www.python.org)
 [![docs](https://img.shields.io/badge/docs-research.st--andrews-blueviolet)](https://research.st-andrews.ac.uk/kinggroup/peaks)
 [![Code Style: Ruff (Black-compatible)](https://img.shields.io/badge/code%20style-ruff-black)](https://docs.astral.sh/ruff/formatter/)
 
-
 `peaks`: **P**ython **E**lectron Spectroscopy **A**nalysis by **K**ing Group **S**t Andrews.
 
 <!-- overview-start -->
+
 `peaks` provides a collection of analysis tools for the loading, processing and visualisation of spectroscopic data, with a core focus on tools for angle-resolved photoemission.
 
 `peaks` is an evolution of the `PyPhoto` package originally developed by Phil King, Brendan Edwards, Tommaso Antonelli, Edgar Abarca Morales, Lewis Hart, and Liam Trzaska from the [King group](https://www.quantummatter.co.uk/king) at the [University of St Andrews](http://www.st-andrews.ac.uk). This version of `peaks` is the result of a major restructuring of the package in 2023-2025 by Brendan Edwards, Phil King, and Shu Mo.
@@ -19,7 +20,9 @@ Contact [pdk6@st-andrews.ac.uk](pdk6@st-andrews.ac.uk).
 <!-- overview-end -->
 
 ## Citation
+
 <!-- citation-start -->
+
 If you use `peaks` in your work, please cite:
 
 *`peaks`: a Python package for analysis of angle-resolved photoemission and related spectroscopies* \
@@ -30,6 +33,7 @@ Edgar Abarca Morales, Lewis Hart, and Liam Trzaska \
 <!-- citation-end -->
 
 <!-- pypi-installation-start -->
+
 ## Installation
 
 `peaks` is registed on [PyPI](https://pypi.org/project/peaks-arpes/) under the name `peaks-arpes`.
@@ -41,19 +45,22 @@ conda create -n peaks python=3.12
 conda activate peaks
 pip install peaks-arpes
 ```
+
 `peaks` will then be installed together with its core dependencies.
 
 ### Optional dependencies
+
 To install optional dependencies, append `\[dep1, dep2, ...\]` to the end of the `pip install ...` command, where `dep` is the name of the dependency. The following options can currently be specified:
 
 - **structure** - required for the use of the `bz` module, for e.g. plotting Brillouin zones on the data;
 - **ML** - required for the use of the machine learning module;
 - **dev** - optional development dependencies, used for e.g. linting the code and installing pre-commit hooks.
-- **docs** - optional dependencies for building local copies of the documentation. 
+- **docs** - optional dependencies for building local copies of the documentation.
 <!-- pypi-installation-end -->
 
 <!-- intel-mac-tip-start -->
-**Intel mac users:** 
+
+**Intel mac users:**
[…]
