### 2026-03-06 847f44b
prepare for v0.4.9 release: sort out version number and changelog; update contributing guide; update package metadata
