### 2026-04-12 #9096 @phrgab
Thanks @arfon, however I believe we have already updated the paper with everything required, and notified that on 2nd Feb in the above comments chain. Can you please confirm if there was something further that we were meant to do? The updated version of the paper is in the `main` branch of our [repo](https://github.com/phrgab/peaks/tree/main)
