### 2026-04-13 #9096 @arfon
Ah, my apologies, I missed that @phrgab. One small request before @phibeck gets going here, your research impact statement claims the software "has underpinned data analysis in more than 10 academic publications from the developers' group." In your paper however, there aren't any citations to those publications. Could you please add some? Additionally, the claim of external adoption ("several groups who are starting to use  peaks as their primary analysis driver") is positive but currently similarly unsupported.
