### 2026-04-13 fdbab8d
add citations for joss paper
