+++ paper/paper.md
@@ -33,7 +33,7 @@ affiliations:
   - name: SUPA, School of Physics and Astronomy, University of St Andrews, St Andrews, KY16 9SS, UK
     index: 1
 
-date: 2 February 2026
+date: 13 April 2026
 bibliography: paper.bib
 ---
 
@@ -65,11 +65,11 @@ In the future, `peaks` could be augmented with additional ML approaches tailored
 
 Being primarily intended for use by experimental physicists, a high-level programming language is required, with Python the natural choice. While the underlying experimental data acquired from the spectrometer is array-like, the varied co-ordinates involved (angles, energies, spatial positions, temperatures, and more) mean that it is crucial to keep this tightly coupled with multiple axis scales and names. The labelled tensor data structure of `xarray` provides an ideal structure for this, and so we utilised this as our fundamental data structure. This has further advantages that it can be used with `dask` arrays, allowing lazy data loading and parallel processing reflecting the increasing data volumes in ARPES. It also provides a data structure that is similar to legacy software in the field, which has conventionally used the software package `Igor Pro`. It also provides excellent integration with interactive notebooks, which provide an ideal platform for use, e.g., during team-based experimental campaigns (as are common in this community) in order to provide a running logbook. Where possible, standard packages were utilised for data analysis (e.g. peak fitting is based around `lmfit`), and for enforcing robust data and metadata structures (e.g. `pydantic` models).
 
-`peaks` was originally developed and tested within the authors' research group. At the time, the only available extensive Python ARPES package that was - to our knowledge - openly available was `PyARPES`, which made a number of fundamentally different choices in regard to the treatment of the experimental data and handling of conventions. Given this, and the design goals of `peaks` discussed above, a choice was made to develop a new package. Since then, some additional packages have been released (see [statement of need](#statement_of_need) section for more details), but with still some different choices made and by when `peaks` was already in a rather developed state.
+`peaks` was originally developed and tested within the authors' research group. At the time, the only available extensive Python ARPES package that was - to our knowledge - openly available was `PyARPES`, which made a number of fundamentally different choices in regard to the treatment of the experimental data and handling of conventions. Given this, and the design goals of `peaks` discussed above, a choice was made to develop a new package. Since then, some additional packages have been released (see [statement of need](#statement-of-need) section for more details), but with still some different choices made and by when `peaks` was already in a rather developed state.
 
 # Research impact statement
 
-The development of `peaks` has so far underpinned data analysis in more than 10 academic publications from the developers’ group during which time it has undergone continuous and active development and maintenance. Since its public release in 2025, we have been contacted by several groups who are starting to use `peaks` as their primary analysis driver for ARPES, and so we expect its impact to substantially broaden over the coming years.
+The development of `peaks` has so far underpinned data analysis in more than 10 academic publications from the developers’ group (e.g. [@buchberger_persistence_2025; @armitage_electronic_2025; @mo_resonant_2026; @saika_flat_2025; @siemann_dichotomy_2025; @edwards_giant_2023]), during which time it has undergone continuous and active development and maintenance. Since its public release in 2025, we have been contacted by several groups who are starting to use `peaks` as their primary analysis driver for ARPES, reflecting increased community adoption (e.g. [@peaks_pr3]), and so we expect its impact to substantially broaden over the coming years.
 
[…]
