### 2026-04-13 #9096 @phrgab
Thanks @arfon, no problem. We've added some citations to relevant papers as requested. For the external adoption to date, most information we have on this is from private communications, but we added a citation/link to a recent PR showing community adoption as evidenced by the creation of a data loader for a new experimental facility previously not covered by the `peaks` loaders. I hope that will be OK. Please let us know if we should do anything further in the mean time.
