### 2026-04-14 #9096 @phibeck
Hi @phrgab my apologies for the delay, and thank you for adding the citations. I'll be looking for reviewers next. When you have a moment it would be helpful if you could identify a few potential reviewers from this list https://reviewers.joss.theoj.org/reviewers (without tagging them with an @), thanks.
