### 2026-04-20 #10440 @jsdodge
As I mentioned in the [pre-review thread](https://github.com/openjournals/joss-reviews/issues/9096#issuecomment-4261238488), I won't be able to start on this until May 1. I'll do my best to complete the review in a timely manner after that.
