### 2026-04-23 d69bcc3
prepare for v0.5.0 release: sort out version number and changelog
