+++ CHANGELOG.md
@@ -5,6 +5,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.5.1.dev]
+
+### Fixed
+
+- Error when loading NetCDF files saved by `peaks` with `metadata=False`.
+
 ## [0.5.0] - 2026-04-23
 
 ### Added
