### 2026-05-01 #10440 @pudeIko
Hi @phibeck ,
I'm mostly done with my review.
I have a few questions/remarks and just wanted to make sure what's the proper way to share them. Can I just write them as a flat list, or is some particular format expected?
