### 2026-05-01 24f4b88
always parse metadata when loading peaks-saved zarr files
