### 2026-05-05 phrgab/peaks#22 @jsdodge
DOC: broken link
In the [Making changes](https://research.st-andrews.ac.uk/kinggroup/peaks/latest/contributing.html#making-changes) section of the Contributing page, the hyperlink to the [linting and formatting conventions](https://research.st-andrews.ac.uk/kinggroup/peaks/latest/contributing.html#linting-and-formatting-with-ruff) is broken.

Also, "installed" is misspelled in the [Stripping tutorial notebooks](https://research.st-andrews.ac.uk/kinggroup/peaks/latest/contributing.html#stripping-tutorial-notebooks) section.

"Available" is misspelled in the [Pre-commit hook](https://research.st-andrews.ac.uk/kinggroup/peaks/latest/contributing.html#pre-commit-hook) section.

Lastly, the [JOSS review checklist](https://github.com/openjournals/joss-reviews/issues/10440#issuecomment-4376405579) includes guidelines on seeking support, which I don't see—unless the link to the King group is the menu bar is intended for this. If so, I suggest making that explicit on the Contributing page.
