### 2026-05-05 77b1c49
fixing readme and docs errors
