### 2026-05-05 5fdc6ea
typo fixes and adding support sections
