+++ CHANGELOG.md
@@ -11,6 +11,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Errors when loading NetCDF files and processing Zarr files saved by `peaks` with `metadata=False` ([PR#20](https://github.com/phrgab/peaks/pull/20))
 
+### Changed
+
+- Minor documentation improvements ([PR#23](https://github.com/phrgab/peaks/pull/23))
+
 ## [0.5.0] - 2026-04-23
 
 ### Added
