### 2026-05-06 53e366f
updating install from source instructions to reference dev branch
