### 2026-05-06 29b67db
adding basic support section to getting started guide
