+++ docs/source/getting_started.md
@@ -1,18 +1,27 @@
 (getting_started)=
+
 # Getting started
+
 ```{include} ../../README.md
 :start-after: <!-- pypi-installation-start -->
 :end-before: <!-- pypi-installation-end -->
 ```
 
-:::{tip} 
+:::{tip}
+
 ```{include} ../../README.md
 :start-after: <!-- intel-mac-tip-start -->
 :end-before: <!-- intel-mac-tip-end -->
 ```
-::: 
+
+:::
 
 ```{include} ../../README.md
 :start-after: <!-- source-installation-start -->
 :end-before: <!-- basic-usage-end -->
+```
+
+```{include} ../../README.md
+:start-after: <!-- source-support-start -->
+:end-before: <!-- basic-support-end -->
 ```
