+++ paper/paper.md
@@ -43,9 +43,11 @@ The electronic band structure, describing the motion and interactions of electro
 
 # Statement of need
 
-Over recent years, significant technological improvements have developed ARPES into a truly multidimensional spectroscopy. Besides the traditional resolution of energy and up to three momentum directions, temperature, spin, spatial, and temporal-dependent ARPES measurements are becoming increasingly common [@king_angle_2021; @sobota_angle-resolved_2021], typically requiring efficient handling and advanced analysis of 3-, 4-, and higher-dimensional datasets. Extensive use of international light sources for performing ARPES measurements, during intensive experiment campaigns running over several days, further motivates a collaborative approach to performing data analysis. There is also an ever-increasing push to incorporate machine learning (ML) methods into the analysis pipeline [@melton_k-means-driven_2020; @kim_deep_2021; @iwasawa_unsupervised_2022; @agustsson_autoencoder_2025], while greater transparency and reproducibility in ARPES data analysis can be ensured by the development and utilisation of open-source packages, with clear and transparent metadata handling [@scheffler_fair_2022].
+Over recent years, significant technological improvements have developed ARPES into a truly multidimensional spectroscopy. Besides the traditional resolution of energy and up to three momentum directions, temperature, spin, spatial, and temporal-dependent ARPES measurements are becoming increasingly common [@king_angle_2021; @sobota_angle-resolved_2021], typically requiring efficient handling and advanced analysis of 3-, 4-, and higher-dimensional datasets. Extensive use of international light sources for performing ARPES measurements, during intensive experiment campaigns running over several days, further motivates a collaborative approach to performing data analysis. There is also an ever-increasing push to incorporate machine learning (ML) methods into the analysis pipeline [@melton_k-means-driven_2020; @kim_deep_2021; @iwasawa_unsupervised_2022; @agustsson_autoencoder_2025], while greater transparency and reproducibility in ARPES data analysis can be ensured by the development and utilisation of open-source packages, with clear and transparent metadata handling [@scheffler_fair_2022]. The above requirements all motivate the use of Python as a modern approach to ARPES data analysis.
 
-The above requirements all motivate the use of Python as a modern approach to ARPES data analysis. To this end, several packages have been developed. `PyARPES` [@stansbury_pyarpes_2020] represents a pioneering development in this direction. It appears to no longer be actively maintained by the original author, although a maintained fork does exist [@pyarpes_fork]. Despite many excellent features, it makes several fundamental convention choices (regarding angular and energy scales and units, alignments, and sign conventions) which, in our view, complicates its use when employed with multiple experimental setups as is typical in the ARPES community, while approximations are used in the critical momentum-space conversions. `pesto` [@pesto] is an excellent easy-to-use alternative, but is heavily oriented towards use with data collected from the Bloch beamline of the Max-IV synchrotron. We have recently discovered `ERLabPy` [@ErLabPy] which provides similar functionality to `peaks`, although with some differences in the approach to handling the data (e.g. co-ordinate systems). The need to accommodate not only different data formats but also manage distinct angle and sign convention choices for data acquired at multiple facilities can add significant complexity for the user, in particular for on-the-fly processing: this is something that `peaks` attempts to simplify for the end user, aiding quick and efficient on-the-fly analysis e.g. for sample alignment during intense experimental runs. Other packages that we are aware of tend to focus on a subset of the functions required, e.g. for ARPES data analysis [@das] or visualistaion [@Kramer2021].
+# State of the field
+
[…]
