### 2026-05-06 phrgab/peaks#22 @phrgab
Thanks @jsdodge for the careful checking of these. These and #21 should now be fixed on the `dev` branch with #23
