### 2026-05-08 ccbcd0e
remove peaks and dask from intersphinx_mapping as peaks is not an external packageand dask is not registered as a module entry
