### 2026-05-08 7d61961
add dask back to intersphinx mapping as submodule entries are registered
