### 2026-05-08 7d129bd
add API cross-references to explanations, minor fixes to contributing guide and deprecate development.md file that is no longer in use
