+++ docs/source/development.md
@@ -1,8 +1,0 @@
-(development)=
-# Development
-```{toctree}
-:maxdepth: 1
-
-contributing
-changelog
-```
