+++ docs/source/contributing.md
@@ -134,10 +134,10 @@ The core user documentation comes from a combination of the tutorials (Jupyter n
 :::{tip}
 We use the [NumPy docstring format](https://numpydoc.readthedocs.io/en/latest/format.html). Specify Parameters and give relevant code examples. To make a code block, insert `::` at the end of the line before, and then leave a space and indent the text.
 
-We are using `intersphinx` to make links to other documentation, which can be quite helpful. But this only works if we put in the full name not the abbreviation for the other package. so use e.g. `xarray.DataArray` and not `xr.DataArray` in type annotations in the docstring. This works out of the box if this is given as the type for a parameter or returns. If you want to include this within part of the general text, need to do `:class:xarray.DataArray`. See other modules for examples.
+We are using `intersphinx` to make links to other documentation, which can be quite helpful. But this only works if we put in the full name not the abbreviation for the other package. so use e.g. `xarray.DataArray` and not `xr.DataArray` in type annotations in the docstring. This works out of the box if this is given as the type for a parameter or returns. If you want to include this within part of the general text, need to do ``{py:class}`xarray.DataArray` ``. See other modules for examples.
 :::
 
-The hosted documentation is automatically built by Gitlab CI, and updated on each merge to the main branch and on the release of a new tagged version. It is important that all of the tutorials can run without any local data files. If specific example data is required which is not already avaialble in the `peaks.core.utils.sample_data` module, raise an Issue to discuss adding a new example dataset there.
+The hosted documentation is automatically built by Gitlab CI, and updated on each merge to the `dev` branch and on the release of a new tagged version. It is important that all of the tutorials can run without any local data files. If specific example data is required which is not already avaialble in the {py:mod}`peaks.core.utils.sample_data` module, raise an Issue to discuss adding a new example dataset there.
 
 :::{tip}
 To make a local build of the documentation, install `peaks` including the optional `[docs]` dependency. To build the documentation, navigate to the `docs` directory and run one of the following commands:
