### 2026-05-08 c9ec625
bump sphinx and myst-parser upper bounds
