+++ pyproject.toml
@@ -81,12 +81,12 @@ structure = [
 ]
 ML = ["scikit-learn (>=1.3,<2)"]
 docs = [
-        "sphinx (>=8.2,<9)",
+        "sphinx (>=8.2,<10)",
         "sphinx-copybutton (>=0.5,<1)",
         "sphinx-togglebutton (>=0.3,<1)",
         "sphinx-inline-tabs (>=2023)",
         "sphinx-autoapi (>=3.6,<4)",
-        "myst-parser (>=4.0,<5)",
+        "myst-parser (>=5.0,<6)",
         "myst-nb (>=1.2,<2)",
         "pydata-sphinx-theme (>=0.16,<1)",
         ]
