### 2026-05-09 952bb8e
add a page for peaks added xr accessors using a sphinx custom extension
