### 2026-05-11 1e61615
generalise accessor generation across dt and ds ones
