+++ docs/source/_ext/list_peaks_accessors.py
@@ -15,7 +15,7 @@ from sphinx.application import Sphinx
 from sphinx.util.typing import ExtensionMetadata
 
 
-# formatters and retrieve accessor info from source
+# formatters and row builders
 def _format_table(rows: list[tuple[str, str]]) -> str:
     header = "| Accessor | Source function |\n"
     line = "|---|---|\n"
@@ -23,16 +23,46 @@ def _format_table(rows: list[tuple[str, str]]) -> str:
     return header + line + body
 
 
-def _dataarray_rows() -> list[tuple[str, str]]:
-    mod = importlib.import_module("peaks.core.accessors.dataarray_accessors")
+def _registry_rows(module_name: str, registry_attr: str) -> list[tuple[str, str]]:
+    mod = importlib.import_module(module_name)
+    registry = getattr(mod, registry_attr)
     rows: list[tuple[str, str]] = []
-    for sub_module, func_names in mod.functions_to_register.items():
+    for sub_module, func_names in registry.items():
         full_module = f"peaks.core.{sub_module}"
         for func_name in func_names:
             rows.append((func_name, f"{full_module}.{func_name}"))
     return rows
 
 
+def _dataarray_rows() -> list[tuple[str, str]]:
+    return _registry_rows(
+        "peaks.core.accessors.dataarray_accessors",
+        "functions_to_register",
+    )
+
+
+def _dataset_rows() -> list[tuple[str, str]]:
+    return _registry_rows(
+        "peaks.core.accessors.dataset_accessors",
+        "functions_to_register",
+    )
+
+
+def _datatree_direct_rows() -> list[tuple[str, str]]:
+    return _registry_rows(
+        "peaks.core.accessors.datatree_accessors",
+        "functions_to_register_for_direct_accessor",
+    )
+
+
+def _datatree_iter_rows() -> list[tuple[str, str]]:
+    return _registry_rows(
+        "peaks.core.accessors.datatree_accessors",
+        "functions_to_register_for_iterable_accessor",
+    )
+
+
+# dedicated row builder for time-resolved accessors
 def _tr_rows() -> list[tuple[str, str]]:
     rows: list[tuple[str, str]] = []
     tr_mod = importlib.import_module("peaks.time_resolved.accessors")
@@ -55,22 +85,55 @@ def _build_page() -> str:
 
         `peaks` registers many of its core functions as
         [xarray accessors](https://docs.xarray.dev/en/latest/internals/extending-xarray.html) which
-        can be called as methods acting directly on an {py:class}`xarray.DataArray`. For example,
-        instead of `peaks.core.process.data_select.MDC(da, E=6.2, dE=0.01)`, `da.MDC(6.2, 0.01)` can be used.
+        can be called as methods acting directly on an {py:class}`xarray.DataArray`, {py:class}`xarray.Dataset`, or
+        {py:class}`xarray.DataTree`. For example, instead of `peaks.core.process.data_select.MDC(da, E=6.2, dE=0.01)`,
+        `da.MDC(6.2, 0.01)` can be used, where `da` is the {py:class}`xarray.DataArray` holding the data loaded by `peaks`.
 
-        ## Direct accessors
+        ## DataArray accessors
 
         Called directly on an {py:class}`xarray.DataArray`, e.g. `da.EDC(...)`.
 
         """
     )
 
-    heading = dedent(
+    dataset_heading = dedent(
+        """\
+
+            ## Dataset accessors
+
+            Called directly on an {py:class}`xarray.Dataset`.
+
+            """
+    )
+
+    datatree_direct_heading = dedent(
+        """\
+
+        ## DataTree accessors (direct)
+
[…]
