### 2026-05-12 3a73be2
added conda and uv to installation instructions
