+++ README.md
@@ -2,7 +2,8 @@
 
 [![git](https://img.shields.io/badge/repo-github-blue?logo=github)](https://github.com/phrgab/peaks)
 [![GitHub Actions Workflow Status](https://img.shields.io/github/actions/workflow/status/phrgab/peaks/test-installs.yml?logo=github-actions&logoColor=white&label=install%20tests)](https://github.com/phrgab/peaks/actions/workflows/test-installs.yml)
-[![Latest version on pypi](https://img.shields.io/pypi/v/peaks-arpes?logo=pypi&logoColor=white&color=brightgree)](https://pypi.org/project/peaks-arpes/)
+[![Latest version on pypi](https://img.shields.io/pypi/v/peaks-arpes?logo=pypi&logoColor=white&color=green)](https://pypi.org/project/peaks-arpes/)
+[![Latest version on conda-forge](https://img.shields.io/conda/vn/conda-forge/peaks-arpes?logo=conda-forge&logoColor=white&color=green)](https://anaconda.org/channels/conda-forge/packages/peaks-arpes/overview)
 [![Supported Python versions](https://img.shields.io/pypi/pyversions/peaks-arpes?logo=python&logoColor=white)](https://www.python.org)
 [![docs](https://img.shields.io/badge/docs-research.st--andrews-blueviolet)](https://research.st-andrews.ac.uk/kinggroup/peaks)
 [![Code Style: Ruff (Black-compatible)](https://img.shields.io/badge/code%20style-ruff-black)](https://docs.astral.sh/ruff/formatter/)
@@ -30,45 +31,74 @@ Edgar Abarca Morales, Lewis Hart, and Liam Trzaska \
 
 <!-- citation-end -->
 
-<!-- pypi-installation-start -->
+<!-- main-installation-start -->
 
 ## Installation
 
-`peaks` is registed on [PyPI](https://pypi.org/project/peaks-arpes/) under the name `peaks-arpes`.
+`peaks` is registered on [PyPI](https://pypi.org/project/peaks-arpes/) under the name `peaks-arpes`, and is also distributed through the [conda-forge](https://anaconda.org/channels/conda-forge/packages/peaks-arpes/overview) channel under the same name `peaks-arpes` where the version may lag slightly behind that on [PyPI](https://pypi.org/project/peaks-arpes/).
 
-`peaks` has been tested with Python 3.11 through 3.14. It is recommended to install `peaks` in its own isolated environment. E.g. using conda:
+`peaks` has been tested with Python 3.11 through 3.14. It is recommended to install `peaks` in its own isolated environment. E.g.
 
-```bash
-conda create -n peaks python=3.12
-conda activate peaks
-pip install peaks-arpes
-```
+- **Using [`uv`](https://docs.astral.sh/uv) with a managed project:**
+
+  ```bash
+  uv init peaks -p 3.13
+  cd peaks
+  uv add peaks-arpes
+  ```
+
+- **Using [`conda`](https://conda.org) with an installer such as [Miniforge](https://conda-forge.org/download/) or [Miniconda](https://docs.anaconda.com/miniconda/):**
+
+  ```bash
+  conda create -n peaks python=3.13
+  conda activate peaks
+  conda install -c conda-forge peaks-arpes # or install from PyPI: pip install peaks-arpes
+  ```
+
+- **Using Python virtual environments** (This works with the default Python version installed on the OS. To use a specific Python version, see, e.g. [uv-managed virtual environments](https://docs.astral.sh/uv/pip/environments/)):
+
+  ```bash
+  python -m venv peaks
+  source peaks/bin/activate  # Windows PowerShell: peaks\Scripts\activate
+  pip install peaks-arpes
+  ```
 
[…]
