+++ docs/source/getting_started.md
@@ -3,15 +3,29 @@
 # Getting started
 
 ```{include} ../../README.md
-:start-after: <!-- pypi-installation-start -->
-:end-before: <!-- pypi-installation-end -->
+:start-after: <!-- main-installation-start -->
+:end-before: <!-- main-installation-end -->
 ```
 
-:::{tip}
+:::{note}
 
 ```{include} ../../README.md
 :start-after: <!-- intel-mac-tip-start -->
 :end-before: <!-- intel-mac-tip-end -->
+```
+
+:::
+
+```{include} ../../README.md
+:start-after: <!-- extras-installation-start -->
+:end-before: <!-- extras-installation-end -->
+```
+
+:::{note}
+
+```{include} ../../README.md
+:start-after: <!-- conda-extras-note-start -->
+:end-before: <!-- conda-extras-note-end -->
 ```
 
 :::
