### 2026-05-12 2d95948
add note for running tutorial notebooks to user guide and fix error when plotting tutorial figs locally
