### 2026-05-12 a72e219
update changelog and fix CI
