### 2026-05-12 d45968a
typo fixes in the paper
