+++ paper/paper.bib
@@ -186,11 +186,12 @@
 }
 
 @misc{ErLabPy,
-	author = {K. Han},
+	author = {Higashihira Han, Kimoon},
 	title = {erlab: A Complete Python Workflow for ARPES Experiments},
 	year = {2026},
 	publisher = {Zenodo},
 	url = {https://doi.org/10.5281/zenodo.18392129},
+	doi = {10.5281/zenodo.18392129},
 }
 
 @article{hoyer2017xarray,
