### 2026-05-18 5065bfd
match ROI outline colour to ROI polygon colour
