### 2026-05-18 b158faa
extend func to embed videos in docs and preserve figure aspect ratios
