+++ docs/source/conf.py
@@ -110,10 +110,12 @@ myst_substitutions = {
     "date": date.today().strftime("%d.%m.%Y"),
 }
 
+html_static_path = ["_static"]  # for custom CSS used below
 html_css_files = [
     "https://cdnjs.cloudflare.com/ajax/libs/font-awesome/6.0.0/css/fontawesome.min.css",
     "https://cdnjs.cloudflare.com/ajax/libs/font-awesome/6.0.0/css/solid.min.css",
     "https://cdnjs.cloudflare.com/ajax/libs/font-awesome/6.0.0/css/brands.min.css",
+    "ensure_ratio.css",
 ]
 
 html_sidebars = {
