+++ CHANGELOG.md
@@ -12,6 +12,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Cross-references to the API documentation throughout the user guide ([PR#26](https://github.com/phrgab/peaks/pull/26))
 - List of accessors that `peaks` adds on top of `xarray` on the explanations page ([PR#26](https://github.com/phrgab/peaks/pull/26))
 - conda-forge package publishing
+- Demo video showing how to use the ROI panel in the 4D GUI
+- `plot_tutorial_example_figure` now renders videos
 
 ### Fixed
 
@@ -22,6 +24,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Minor documentation improvements ([PR#23](https://github.com/phrgab/peaks/pull/23))
 - Installation instructions now cover `uv`, conda-forge and the traditional Python venv approach ([PR#28](https://github.com/phrgab/peaks/pull/28))
+- Figure aspect ratios in output cells are preserved on docs build
 
 ## [0.5.0] - 2026-04-23
 
