### 2026-05-18 eec3762
fix empty iplot of SM1 in user guide
