+++ peaks/core/GUI/iplot/hvplot.py
@@ -1,5 +1,7 @@
 """Thin wrapper around hvplot to set some defaults and automate handling pint dequantify"""
 
+import os
+
 import xarray as xr
 
 
@@ -14,5 +16,11 @@ class HVPlotAccessor:
 
     def __call__(self, *args, **kwargs):
         # Generate the plot
+        if (
+            os.getenv("FORCE_NB_EXECUTION") == "1"
+        ):  # 1: in a docs build see build_docs.sh
+            kwargs.setdefault(
+                "dynamic", False
+            )  # force static plot to avoid issues with dynamic plots but locally keep dynamic=True (default) for performance
         plot = self._obj.pint.dequantify().hvplot(*args, **kwargs)
         return plot
