### 2026-05-19 90da919
auto cosmetic fixes for numpy-style docstrings
