+++ peaks/core/GUI/disp_panels/degrid_panel.py
@@ -1,7 +1,0 @@
-"""Functions for the mesh grid remover display panel."""
-
-# Brendan Edwards 07/11/2023
-
-
-def degrid():
-    raise NotImplementedError("This UI is not yet implemented.")

[…]
