### 2026-05-19 a7e3420
enable D101 D103 to enforce docstrings on classes and fucntions but exempt docs/ and tests/; bump ruff target v from py312 to py313
