+++ pyproject.toml
@@ -94,13 +94,20 @@ dev = ["ruff","pre-commit","nbstripout","pytest"]
 
 [tool.ruff]
 line-length = 89
-target-version = "py312"
+target-version = "py313"
 exclude = ["*.ipynb"]
 
 [tool.ruff.lint]
-select = ["E", "F", "I", "B"]
+select = ["E", "F", "I", "B", "D101", "D103"]
 ignore = ["E203", "E501"]
 fixable = ["I"]
+
+[tool.ruff.lint.per-file-ignores]
+"docs/**" = ["D101", "D103"]
+"tests/**" = ["D101", "D103"]
+
+[tool.ruff.lint.pydocstyle]
+convention = "numpy"
 
 [tool.flit.module]
 name = "peaks"
