### 2026-05-19 54e0497
add missing docstrings
