+++ peaks/core/fileIO/base_data_classes/base_photon_source_classes.py
@@ -58,6 +58,8 @@ class BasePhotonSourceDataLoader(BaseDataLoader):
 
 
 class BasePumpProbeClass(BasePhotonSourceDataLoader):
+    """Mix-in adding pump-photon metadata (hv, polarisation, delay, power, to) to a loader."""
+
     _pump_photon_attributes = ["hv", "polarisation", "delay", "power"]
     _pump_photon_exclude_from_metadata_warn = _pump_photon_attributes
     _metadata_parsers = ["_parse_pump_photon_metadata"]
