### 2026-05-20 8b48fe0
adding missing docstrings
