### 2026-05-21 c37d53b
support multiple disp panels if event loop in ipython in enabled for qt
