+++ peaks/core/GUI/disp_panels/disp_2d.py
@@ -25,6 +25,8 @@ from peaks.core.GUI.GUI_utils.cursor_stats import _parse_norm_emission_cursor_st
 from peaks.core.metadata.metadata_methods import display_metadata
 from peaks.core.process.tools import estimate_sym_point, sym
 
+_active_viewers: list = []  # to store active viewer instances and prevent garbage collection
+
 
 def _disp_2d(data, primary_dim, exclude_from_centering):
     """Display a 2D interactive display panel.
@@ -47,8 +49,41 @@ def _disp_2d(data, primary_dim, exclude_from_centering):
         app = QApplication(sys.argv)
 
     viewer = _Disp2D(data, primary_dim, exclude_from_centering)
+    _active_viewers.append(viewer)  # add the viewer to active viewers list
+    viewer.destroyed.connect(
+        lambda *_: _active_viewers.remove(viewer)
+        if viewer in _active_viewers
+        else None  # Remove viewer from active viewers list when it is closed
+    )
     viewer.show()
-    app.exec()
+
+    # try:
+    #     from IPython import get_ipython
+
+    #     ip = get_ipython()
+    #     print(f"ipython event loop is {ip.active_eventloop}")
+    #     if ip is not None:
+    #         if getattr(ip, "active_eventloop", None) is None:
+    #             ip.enable_gui("qt6")
+    #             print(f"ipython event loop integration is enabled for qt: {ip.active_eventloop}")
+    #         return
+    # except Exception:
+    #     pass
+
+    # to support multiple display panels
+    try:
+        from IPython import get_ipython
+
+        ip = get_ipython()
+        if ip is not None:
+            if getattr(ip, "active_eventloop", None) == "qt6":
+                return
+    except Exception:
+        pass
+
+    # fallback to Qt event loop if not in IPython or if IPYthon does not have an active event loop
+    if not any(v.isVisible() for v in _active_viewers if v is not viewer):
+        app.exec()
 
 
 class _Disp2D(QtWidgets.QMainWindow):
@@ -121,9 +156,9 @@ class _Disp2D(QtWidgets.QMainWindow):
         self.closeEvent = self._close_application
 
     def _close_application(self, event):
-        """Close the application when the window is closed."""
+        """Close the application when the window is closed without shutting down the event loop."""
         self.graphics_layout.close()
-        app.quit()
+        event.accept()
 
     # ##############################
     # GUI layout
