+++ peaks/core/GUI/disp_panels/disp_4d.py
@@ -22,6 +22,8 @@ from ..GUI_utils import (
 )
 from .disp_2d import _Disp2D
 
+_active_viewers: list = []  # to store active viewer instances and prevent garbage collection
+
 
 def _disp_4d(data, primary_dim):
     """Display a 4D interactive display panel.
@@ -41,8 +43,26 @@ def _disp_4d(data, primary_dim):
         app = QApplication(sys.argv)
 
     viewer = _Disp4D(data, primary_dim)
+    _active_viewers.append(viewer)
+    viewer.destroyed.connect(
+        lambda *_: _active_viewers.remove(viewer) if viewer in _active_viewers else None
+    )
     viewer.show()
-    app.exec()
+
+    # to support multiple display panels
+    try:
+        from IPython import get_ipython
+
+        ip = get_ipython()
+        if ip is not None:
+            if getattr(ip, "active_eventloop", None) == "qt6":
+                return
+    except Exception:
+        pass
+
+    # fallback to Qt event loop if not in IPython or if IPYthon does not have an active event loop
+    if not any(v.isVisible() for v in _active_viewers if v is not viewer):
+        app.exec()
 
 
 class _Disp4D(QtWidgets.QMainWindow):
@@ -59,6 +79,11 @@ class _Disp4D(QtWidgets.QMainWindow):
 
         # Set up the GUI
         self._init_UI()  # Initialize the layout
+
+        # clean up the widget on close, but don't quit the global app.
+        # https://doc.qt.io/qt-6/qt.html#WidgetAttribute-enum
+        self.setAttribute(QtCore.Qt.WidgetAttribute.WA_DeleteOnClose)
+        self.closeEvent = lambda event: event.accept()
 
     # ##############################
     # GUI layout
