### 2026-05-21 59e1848
document primary_dim param for disp panel in user guide
