### 2026-05-21 564d69a
add D102 to ruff rule list to enforce docstrings on public methods
