+++ peaks/core/fileIO/base_arpes_data_classes/base_arpes_data_class.py
@@ -134,6 +134,7 @@ class BaseARPESDataLoader(
 
     @property
     def analyser_attributes(self):
+        """Return the list of analyser attribute names recognised by the loader."""
         return self._analyser_attributes
 
     @classmethod
