### 2026-05-21 3961b2a
add D104 to ruff rules to enforce docstrings on public packages
