+++ peaks/core/metadata/__init__.py
@@ -1,0 +1,1 @@
+"""Pydantic-based metadata models, analysis history tracking and the `.metadata` xarray accessor."""
