+++ peaks/core/fileIO/base_arpes_data_classes/__init__.py
@@ -1,0 +1,1 @@
+"""Base data loader classes for ARPES-specific data formats and analyser/manipulator metadata."""
