+++ CHANGELOG.md
@@ -13,7 +13,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - List of accessors that `peaks` adds on top of `xarray` on the explanations page ([PR#26](https://github.com/phrgab/peaks/pull/26))
 - conda-forge package publishing
 - Demo video showing how to use the ROI panel in the 4D GUI ([PR#30](https://github.com/phrgab/peaks/pull/30))
-- D101 and D103 to `ruff` linting rules to enforce docstrings on public classes and functions ([PR#30](https://github.com/phrgab/peaks/pull/30))
+- D101, D102, D103 and D104 to `ruff` linting rules to enforce docstrings on public classes, methods, functions and packages ([PR#30](https://github.com/phrgab/peaks/pull/30))
 - numpy convention to `ruff` linting ([PR#30](https://github.com/phrgab/peaks/pull/30))
 
 ### Fixed
