+++ pyproject.toml
@@ -98,7 +98,7 @@ target-version = "py313"
 exclude = ["*.ipynb"]
 
 [tool.ruff.lint]
-select = ["E", "F", "I", "B", "D101", "D102", "D103"]
+select = ["E", "F", "I", "B", "D101", "D102", "D103", "D104"]
 ignore = ["E203", "E501"]
 fixable = ["I"]
 
