### 2026-05-22 f4469ca
updated loader discussion in docs
