### 2026-05-22 915a011
minor text improvements
