### 2026-05-22 bb9c772
wee changes to correct the example in disp docstring
