### 2026-05-22 b046e18
firing a warning if there are already 3 or more disp panels open
