+++ peaks/core/GUI/disp_panels/disp_4d.py
@@ -16,13 +16,13 @@ from PyQt6.QtWidgets import (
     QWidget,
 )
 
+from peaks.core.utils.misc import analysis_warning
+
 from ..GUI_utils import (
     Crosshair,
     KeyPressGraphicsLayoutWidget,
 )
 from .disp_2d import _Disp2D
-
-_active_viewers: list = []  # to store active viewer instances and prevent garbage collection
 
 
 def _disp_4d(data, primary_dim):
@@ -42,10 +42,24 @@ def _disp_4d(data, primary_dim):
     if app is None:
         app = QApplication(sys.argv)
 
+    if not hasattr(app, "_peaks_active_viewers"):
+        app._peaks_active_viewers = []
+    active_viewers = app._peaks_active_viewers
+
     viewer = _Disp4D(data, primary_dim)
-    _active_viewers.append(viewer)
+    active_viewers.append(viewer)
+
+    # fire a warning if there are already 3 or more disp panels open
+    if len(active_viewers) >= 3:
+        analysis_warning(
+            f"There are currenly {len(active_viewers)} active display panels. "
+            "This may cause performance issues.",
+            warn_type="warning",
+            title="Multiple dispaly panels open",
+        )
+
     viewer.destroyed.connect(
-        lambda *_: _active_viewers.remove(viewer) if viewer in _active_viewers else None
+        lambda *_: active_viewers.remove(viewer) if viewer in active_viewers else None
     )
     viewer.show()
 
@@ -61,7 +75,7 @@ def _disp_4d(data, primary_dim):
         pass
 
     # fallback to Qt event loop if not in IPython or if IPYthon does not have an active event loop
-    if not any(v.isVisible() for v in _active_viewers if v is not viewer):
+    if not any(v.isVisible() for v in active_viewers if v is not viewer):
         app.exec()
 
 
@@ -80,8 +94,7 @@ class _Disp4D(QtWidgets.QMainWindow):
         # Set up the GUI
         self._init_UI()  # Initialize the layout
 
-        # clean up the widget on close, but don't quit the global app.
-        # https://doc.qt.io/qt-6/qt.html#WidgetAttribute-enum
+        # Clean up the widget on close, but don't quit the global app.
         self.setAttribute(QtCore.Qt.WidgetAttribute.WA_DeleteOnClose)
         self.closeEvent = lambda event: event.accept()
 
