### 2026-05-22 0b61698
updated docstrings for loaders
