+++ peaks/core/fileIO/data_loading.py
@@ -23,7 +23,7 @@ def load(
     quiet=False,
     **kwargs,
 ):
-    """Core function to load data.
+    """Load one or more files into :mod:`xarray` objects using the registered loaders.
 
     Parameters
     ------------
@@ -67,25 +67,27 @@ def load(
 
     Returns
     ------------
-    loaded_data : xarray.DataArray, xarray.DataSet, xarray.DataTree
-        The loaded data.
+    loaded_data : xarray.DataArray, xarray.Dataset, xarray.DataTree
+        The loaded object. A single file usually returns a :class:`xarray.DataArray`
+        or :class:`xarray.Dataset`; multiple files are combined into a
+        :class:`xarray.DataTree`.
 
     Notes
     ------------
     Much of file path can be set by :class:`peaks.core.options.FileIO` global options:
 
-        opts.FileIO.path : str, list
+        opts.FileIO.path : :class:`str`, :class:`list`
             Path (or list of paths) to folder(s) where data is stored,
-            e.g. `opts.FileIO.path = 'C:/User/Documents/i05-1-123'`.
+            e.g. ``opts.FileIO.path = 'C:/User/Documents/i05-1-123'``.
 
-        opts.FileIO.ext : str, list
-            Extension (or list of) of data, e.g. `opts.FileIO.ext = ['ibw', 'zip']`.
+        opts.FileIO.ext : :class:`str`, :class:`list`
+            Extension (or list of) of data, e.g. ``opts.FileIO.ext = ['ibw', 'zip']``.
 
-        opts.FileIO.loc : str
-            Location identifier for where data was acquired, e.g. `opts.FileIO.loc = 'MAX IV Bloch'`.
-            Current supported options can be obtained using `pks.locs()`.
+        opts.FileIO.loc : :class:`str`
+            Location identifier for where data was acquired, e.g. ``opts.FileIO.loc = 'MAX IV Bloch'``.
+            Current supported options can be obtained using :func:`peaks.locs`.
 
-        opts.FileIO.lazy_size : int
+        opts.FileIO.lazy_size : :class:`int`
             Size in bytes above which data is loaded in a lazily evaluated dask format.
             Defaults to 1 GB.
 
@@ -113,18 +115,18 @@ def load(
 
         # Load data without needing to define data path or extension, nor the repeated
         # part of the scan name
-        disp3 = load(456)
-        disp4 = load(457)
+        disp3 = pks.load(456)
+        disp4 = pks.load(457)
 
         # Provide part of the scan name using a glob character
-        disps = load('45*')
+        disps = pks.load('45*')
 
         # Load multiple files at once
-        disps = load([456, 457, 458])
+        disps = pks.load([456, 457, 458])
 
         # Still can load data using a complete data path
         # global options defined in `pks.opts.FileIO` will be ignored
-        disp1 = load('C:/User/Documents/Data/disp1.ibw')
+        disp1 = pks.load('C:/User/Documents/Data/disp1.ibw')
 
         # Load data in a lazily evaluated dask format
         disp1 = load('C:/User/Documents/Data/disp1.ibw', lazy=True)
