+++ peaks/core/fileIO/base_arpes_data_classes/base_mbs_class.py
@@ -15,19 +15,16 @@ ureg = pint_xarray.unit_registry
 
 @register_loader
 class MBSDataLoader(BaseARPESDataLoader):
-    """Generic data loader for MBS data.
+    """Generic ARPES loader for MBS A1 spectrometers.
 
-    Notes
-    ------------
-    This class is intended to be subclassed to provide specific loaders for different locs using MBS data formats.
+    Supports the .txt and .krx formats used by MBS systems. This class is intended
+    to be subclassed for specific instruments which should then define local axis
+    conventions and any metadata remapping required for site-specific keys.
 
-    Subclasses should define the `_MBS_metadata_key_mappings` class variable to map any custom metadata fields or
-    fixed values to the standard peaks metadata keys. See the docstrings for `_MBS_metadata_dict_keys_to_peaks_keys`
-    for more information.
-
-    Subclasses should also define the `_MBS_metadata_units` dictionary to map any fixed units that are not determined
-    as part of the metadata loader.
-
+    ``_MBS_metadata_key_mappings`` maps MBS metadata fields or fixed values onto the
+    standard :mod:`peaks` raw metadata keys consumed by ``_load_metadata()`` and the base
+    metadata parsers. ``_MBS_metadata_units`` can be used to supply fixed units when
+    they are not recoverable from the file metadata contents.
     """
 
     _loc_name = "MBS"
@@ -360,7 +357,7 @@ class MBSDataLoader(BaseARPESDataLoader):
     @staticmethod
     def _MBS_metadata_to_dict_w_MBS_keys(metadata_lines):
         """Convert metadata lines to a dictionary of key-value pairs with the keys being the metadata entries
-        as they appear in the MBS metadata (i.e. not yet in :class:`peaks` convention).
+        as they appear in the MBS metadata (i.e. not yet in :mod:`peaks` convention).
 
         Returns
         ------------

[…]
