### 2026-05-22 7239420
small consistency improvements
