+++ peaks/core/GUI/disp_panels/disp_4d.py
@@ -52,10 +52,10 @@ def _disp_4d(data, primary_dim):
     # fire a warning if there are already 3 or more disp panels open
     if len(active_viewers) >= 3:
         analysis_warning(
-            f"There are currenly {len(active_viewers)} active display panels. "
+            f"There are currently {len(active_viewers)} active display panels. "
             "This may cause performance issues.",
             warn_type="warning",
-            title="Multiple dispaly panels open",
+            title="Multiple display panels open",
         )
 
     viewer.destroyed.connect(
