### 2026-05-24 944371b
initial isolated dead pixel correction
