+++ CHANGELOG.md
@@ -15,6 +15,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Demo video showing how to use the ROI panel in the 4D GUI ([PR#30](https://github.com/phrgab/peaks/pull/30))
 - D101 and D103 to `ruff` linting rules to enforce docstrings on public classes and functions ([PR#30](https://github.com/phrgab/peaks/pull/30))
 - numpy convention to `ruff` linting ([PR#30](https://github.com/phrgab/peaks/pull/30))
+- support for opening multiple `disp` panels simultaneously in Jupyter environments when `%gui qt6` is enabled ([PR#34](https://github.com/phrgab/peaks/pull/34))
 
 ### Fixed
 
