### 2026-05-25 b85eb45
syntax fix in examples
