### 2026-05-25 2b43235
linking hdf5_explorer into developer docs
Adding tip for hdf5_explorer from BaseHDF5DataLoader class into
developer docs.
Updating hdf5_explorer docstring for consistency
