### 2026-05-26 79afbc3
add unit tests for metadata displya
