+++ tests/core/metadata/test_metadata_methods.py
@@ -1,0 +1,131 @@
+import pytest
+
+from peaks.core.metadata.metadata_methods import (
+    DARK_BG_PALETTE,
+    LIGHT_BG_PALETTE,
+    display_metadata,
+)
+from peaks.core.utils.sample_data import ExampleData
+
+
+@pytest.fixture(scope="session")
+def disp():
+    return ExampleData.dispersion2a()
+
+
+# for testing ANSI output
+# copied from the internal _hex_to_rgb function in metadata_methods
+def _hex_to_rgb(hex_col):
+    hex_col = hex_col.lstrip("#")
+    return tuple(int(hex_col[i : i + 2], 16) for i in (0, 2, 4))
+
+
+class TestDisplayMetadata:
+    def test_colours_applied_to_nesting_levels_html(self, disp):
+        result = display_metadata(disp, mode="HTML")
+        expected_pairs = [
+            (0, "calibration"),
+            (1, "EF_correction"),
+            (2, "lens_mode"),
+            (3, "width"),
+        ]
+        for level, key in expected_pairs:
+            expected = f"<span style='color:{LIGHT_BG_PALETTE[level]}'>{key}:</span>"
+            assert expected in result
+
+    def test_colours_applied_to_nesting_levels_html_dark(self, disp):
+        result = display_metadata(disp, palette=DARK_BG_PALETTE, mode="HTML")
+        expected_pairs = [
+            (0, "calibration"),
+            (1, "EF_correction"),
+            (2, "lens_mode"),
+            (3, "width"),
+        ]
+        for level, key in expected_pairs:
+            expected = f"<span style='color:{DARK_BG_PALETTE[level]}'>{key}:</span>"
+            assert expected in result
+
+    def test_colour_cycling_wraps(self):
+        deep_nesting = {"L0": {"L1": {"L2": {"L3": {"L4": "value"}}}}}
+        result = display_metadata(deep_nesting, mode="HTML")
+        assert result.count(f"color:{LIGHT_BG_PALETTE[0]}") == 2
+        for i in (1, 2, 3):
+            assert result.count(f"color:{LIGHT_BG_PALETTE[i]}") == 1
+        assert f"<span style='color:{LIGHT_BG_PALETTE[0]}'>L0:</span>" in result
+        assert f"<span style='color:{LIGHT_BG_PALETTE[0]}'>L4:</span>" in result
+
+    def test_invalid_mode_raises(self, disp):
+        with pytest.raises(ValueError, match="mode must be"):
+            display_metadata(disp, mode="och")
+
+    def test_colours_applied_to_nesting_levels_ansi(self, disp):
+        result = display_metadata(disp, mode="ANSI")
+        expected_pairs = [
+            (0, "calibration"),
+            (1, "EF_correction"),
+            (2, "lens_mode"),
+            (3, "width"),
+        ]
+        for level, key in expected_pairs:
+            r, g, b = _hex_to_rgb(LIGHT_BG_PALETTE[level])
+            expected = f"\x1b[38;2;{r};{g};{b}m{key}\x1b[0m"
+            assert expected in result
+
+    def test_colours_applied_to_nesting_levels_ansi_dark(self, disp):
+        result = display_metadata(disp, palette=DARK_BG_PALETTE, mode="ANSI")
+        expected_pairs = [
+            (0, "calibration"),
+            (1, "EF_correction"),
+            (2, "lens_mode"),
+            (3, "width"),
+        ]
+        for level, key in expected_pairs:
+            r, g, b = _hex_to_rgb(DARK_BG_PALETTE[level])
+            expected = f"\x1b[38;2;{r};{g};{b}m{key}\x1b[0m"
+            assert expected in result
+
+    def test_colour_cycling_wraps_ansi(self):
[…]
