### 2026-05-26 428c29b
use synthetic dictionary for display_metadata tests
