+++ tests/core/metadata/test_metadata_methods.py
@@ -13,6 +13,11 @@ def disp():
     return ExampleData.dispersion2a()
 
 
+@pytest.fixture
+def test_dict():
+    return {"a": {"b": {"c": {"d": {"e": "value"}}}}}
+
+
 # for testing ANSI output
 # copied from the internal _hex_to_rgb function in metadata_methods
 def _hex_to_rgb(hex_col):
@@ -21,72 +26,70 @@ def _hex_to_rgb(hex_col):
 
 
 class TestDisplayMetadata:
-    def test_colours_applied_to_nesting_levels_html(self, disp):
-        result = display_metadata(disp, mode="HTML")
+    def test_colours_applied_to_nesting_levels_html(self, test_dict):
+        result = display_metadata(test_dict, mode="HTML")
         expected_pairs = [
-            (0, "calibration"),
-            (1, "EF_correction"),
-            (2, "lens_mode"),
-            (3, "width"),
+            (0, "a"),
+            (1, "b"),
+            (2, "c"),
+            (3, "d"),
         ]
         for level, key in expected_pairs:
             expected = f"<span style='color:{LIGHT_BG_PALETTE[level]}'>{key}:</span>"
             assert expected in result
 
-    def test_colours_applied_to_nesting_levels_html_dark(self, disp):
-        result = display_metadata(disp, palette=DARK_BG_PALETTE, mode="HTML")
+    def test_colours_applied_to_nesting_levels_html_dark(self, test_dict):
+        result = display_metadata(test_dict, palette=DARK_BG_PALETTE, mode="HTML")
         expected_pairs = [
-            (0, "calibration"),
-            (1, "EF_correction"),
-            (2, "lens_mode"),
-            (3, "width"),
+            (0, "a"),
+            (1, "b"),
+            (2, "c"),
+            (3, "d"),
         ]
         for level, key in expected_pairs:
             expected = f"<span style='color:{DARK_BG_PALETTE[level]}'>{key}:</span>"
             assert expected in result
 
-    def test_colour_cycling_wraps(self):
-        deep_nesting = {"L0": {"L1": {"L2": {"L3": {"L4": "value"}}}}}
-        result = display_metadata(deep_nesting, mode="HTML")
+    def test_colour_cycling_wraps(self, test_dict):
+        result = display_metadata(test_dict, mode="HTML")
         assert result.count(f"color:{LIGHT_BG_PALETTE[0]}") == 2
         for i in (1, 2, 3):
             assert result.count(f"color:{LIGHT_BG_PALETTE[i]}") == 1
-        assert f"<span style='color:{LIGHT_BG_PALETTE[0]}'>L0:</span>" in result
-        assert f"<span style='color:{LIGHT_BG_PALETTE[0]}'>L4:</span>" in result
+        assert f"<span style='color:{LIGHT_BG_PALETTE[0]}'>a:</span>" in result
+        assert f"<span style='color:{LIGHT_BG_PALETTE[0]}'>e:</span>" in result
 
     def test_invalid_mode_raises(self, disp):
         with pytest.raises(ValueError, match="mode must be"):
             display_metadata(disp, mode="och")
 
-    def test_colours_applied_to_nesting_levels_ansi(self, disp):
-        result = display_metadata(disp, mode="ANSI")
+    def test_colours_applied_to_nesting_levels_ansi(self, test_dict):
+        result = display_metadata(test_dict, mode="ANSI")
         expected_pairs = [
-            (0, "calibration"),
-            (1, "EF_correction"),
-            (2, "lens_mode"),
-            (3, "width"),
+            (0, "a"),
+            (1, "b"),
+            (2, "c"),
+            (3, "d"),
         ]
         for level, key in expected_pairs:
[…]
