### 2026-05-26 989c3ef
add swept hot-pixel correction logic
