+++ peaks/core/process/tools.py
@@ -14,7 +14,12 @@ from peaks.core.fitting.models import _shirley_bg
 from peaks.core.metadata.metadata_methods import compare_metadata
 from peaks.core.process.fermi_level_correction import _flatten_EF
 from peaks.core.utils.datatree_utils import get_list_of_DataArrays_from_DataTree
-from peaks.core.utils.interpolation import _fast_bilinear_interpolate_rectilinear
+from peaks.core.utils.interpolation import (
+    _fast_bilinear_interpolate_rectilinear,
+    _fast_linear_interpolate,
+    _fast_linear_interpolate_rectilinear,
+    _is_linearly_spaced,
+)
 from peaks.core.utils.misc import analysis_warning, dequantify_quantify_wrapper
 
 ureg = pint_xarray.unit_registry
@@ -2009,7 +2014,8 @@ def _make_bad_pixel_mask(data, bad_pixels):
         - Sequence of raw integer index pairs in ``data.dims`` order.
         - Sequence of coordinate dictionaries.
 
-        Coordinate values are matched to the nearest available pixel.
+        Coordinate values are matched using pixel boundaries, treating the
+        stored coordinates as pixel centres.
 
     Returns
     -------
@@ -2033,9 +2039,9 @@ def _make_bad_pixel_mask(data, bad_pixels):
 
     for pixel in bad_pixels:
         if isinstance(pixel, dict):
-            # Find the nearest integer index along each named coordinate.
+            # Map each coordinate value to the pixel whose centre interval contains it.
             idx = {
-                dim: int(np.argmin(np.abs(data[dim].data - value)))
+                dim: _coord_to_pixel_index(data[dim].data, value)
                 for dim, value in pixel.items()
             }
 
@@ -2049,6 +2055,188 @@ def _make_bad_pixel_mask(data, bad_pixels):
         bad.data[i0, i1] = True
 
     return bad
+
+
+def _coord_to_pixel_index(coords, value):
+    """
+    Convert a coordinate value to a pixel index using pixel boundaries.
+
+    The coordinates are interpreted as pixel centres to match display panel,
+    so the selected pixel is the one whose half-step interval contains ``value``.
+    """
+    coords = np.asarray(coords)
+
+    if coords.ndim != 1:
+        raise ValueError("coords must be 1-dimensional")
+
+    if coords.size == 1:
+        return 0
+
+    if coords[0] > coords[-1]:
+        ascending_coords = coords[::-1]
+        edges = np.empty(ascending_coords.size + 1, dtype=float)
+        edges[1:-1] = 0.5 * (ascending_coords[:-1] + ascending_coords[1:])
+        edges[0] = ascending_coords[0] - 0.5 * (
+            ascending_coords[1] - ascending_coords[0]
+        )
+        edges[-1] = ascending_coords[-1] + 0.5 * (
+            ascending_coords[-1] - ascending_coords[-2]
+        )
+        idx = int(
+            np.clip(
+                np.searchsorted(edges, value, side="left") - 1,
+                0,
+                ascending_coords.size - 1,
+            )
+        )
+        return coords.size - 1 - idx
+
+    edges = np.empty(coords.size + 1, dtype=float)
+    edges[1:-1] = 0.5 * (coords[:-1] + coords[1:])
+    edges[0] = coords[0] - 0.5 * (coords[1] - coords[0])
+    edges[-1] = coords[-1] + 0.5 * (coords[-1] - coords[-2])
+
+    return int(
+        np.clip(np.searchsorted(edges, value, side="right") - 1, 0, coords.size - 1)
+    )
+
+
+def _round_half_up(value):
+    """Round half-integers up for positive pixel-index values."""
[…]
