### 2026-05-26 33f00ca
adding unit tests for bad pixel correction
