### 2026-05-26 2530d9a
adding bad pixel correction as accessors
