### 2026-05-26 dc42001
update changelog, readme and contributing guide - dev branch is going away soon
