+++ CHANGELOG.md
@@ -15,6 +15,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Demo video showing how to use the ROI panel in the 4D GUI ([PR#30](https://github.com/phrgab/peaks/pull/30))
 - D101, D102, D103 and D104 to `ruff` linting rules to enforce docstrings on public classes, methods, functions and packages ([PR#30](https://github.com/phrgab/peaks/pull/30))
 - numpy convention to `ruff` linting ([PR#30](https://github.com/phrgab/peaks/pull/30))
+- Helper methods to correct dead or hot pixels in dispersions ([PR#36](https://github.com/phrgab/peaks/pull/36))
 
 ### Fixed
 
