### 2026-05-26 52697e5
modify dt iteration
For dts which contain leaves with only a single da, change how the
da-expecting function is applied to preserve updated metadata like
analyis history
