### 2026-05-26 5915d25
adding unit test to ensure analysis history update for dt methods
