### 2026-05-28 8f8e419
docstring improvement
Co-authored-by: Shu <148200713+shum23@users.noreply.github.com>
