### 2026-05-28 82ca895
docstring improvement
Co-authored-by: Shu <148200713+shum23@users.noreply.github.com>
