### 2026-05-28 3a7c526
docstring improvement
Co-authored-by: Shu <148200713+shum23@users.noreply.github.com>
