### 2026-05-28 54300e3
docstring improvement
Co-authored-by: Shu <148200713+shum23@users.noreply.github.com>
