### 2026-05-28 45ac8fa
docstring improvement
Co-authored-by: Shu <148200713+shum23@users.noreply.github.com>
