### 2026-05-28 daf8cb1
docstring improvement
Co-authored-by: Shu <148200713+shum23@users.noreply.github.com>
