### 2026-05-28 17bb0ae
docstring improvement
Co-authored-by: Shu <148200713+shum23@users.noreply.github.com>
