### 2026-05-28 01eb345
guarding pixel masks larger than 2x
