### 2026-05-28 9889206
using pytest fixture for consistency
