### 2026-05-28 fa68c74
minor wording tweaks
