### 2026-05-29 cb2346c
add the CI status badge
