### 2026-05-29 83e3f13
bump main version to 0.5.2.dev
