### 2026-06-01 d6818f3
fix loc identification reading metadata in wrong format
