+++ peaks/core/fileIO/loc_registry.py
@@ -102,7 +102,9 @@ class IdentifyLoc:
             )
 
             # Extract metadata dictionary from file
-            metadata = BaseDataLoader.load_metadata(fname, loc="SES", quiet=True)
+            metadata = BaseDataLoader.load_metadata(
+                fname, loc="SES", return_as_dict=True, quiet=True
+            )
             location_identifier = metadata.get("local_location_identifier")
             if (
                 "bloch" in location_identifier.lower()
@@ -131,7 +133,11 @@ class IdentifyLoc:
         from peaks.core.fileIO.base_data_classes.base_data_class import BaseDataLoader
 
         # Extract metadata dictionary from file
-        metadata = BaseDataLoader.load_metadata(fname, loc="SES", quiet=True)
+        metadata = BaseDataLoader.load_metadata(
+            fname, loc="SES", return_as_dict=True, quiet=True
+        )
+        print(f"keys are {sorted(metadata.keys())}")
+        print(f"local loc identifier is {metadata.get('local_location_identifier')}")
         location_identifier = metadata.get("local_location_identifier")
         if (
             "bloch" in location_identifier.lower()
