### 2026-06-02 99729ca
test save-load roundtrip preserves metadata
