### 2026-06-02 59f1887
return in plain format rather than pretty format for pint Quantity when copying norm emis from GUI
