### 2026-06-19 9379308
close open hdf5 file handles during sample data cleanup
