### 2026-06-19 d25bda8
parse loader filenames with os.path for win compatibility
