### 2026-06-19 844cf16
close temp CIF handle before loading
