### 2026-06-19 55fdf8b
removing debug lines
