### 2026-06-30 f2ceba2
update the nano bit in the i05 loader following the analyser upgrade
