+++ peaks/core/fileIO/loaders/diamond.py
@@ -506,8 +506,7 @@ class I05NanoARPESLoader(I05ARPESLoader, BaseOpticsDataLoader):
     }
 
     _analyser_sign_conventions = {
-        "theta_par": -1,
-        "deflector_parallel": -1,  # consistent with theta_par
+        # overriden when loading data to take account of analyser upgrade see below
     }
 
     _desired_dim_order = [
@@ -530,8 +529,7 @@ class I05NanoARPESLoader(I05ARPESLoader, BaseOpticsDataLoader):
     ]
 
     _analyser_name_conventions = {
-        "deflector_perp": "ThetaY",
-        "deflector_parallel": "ThetaX",
+        # deflector conventions overridden when loading data to take account of analyser upgrade see below
         "eV": ["energies", "kinetic_energy_center"],
         "theta_par": "angles",
         "hv": ["energy", "value"],
@@ -558,14 +556,61 @@ class I05NanoARPESLoader(I05ARPESLoader, BaseOpticsDataLoader):
         "manipulator_x2": "entry1/instrument/manipulator/smy",
         "manipulator_x3": "entry1/instrument/manipulator/smz",
         "manipulator_defocus": "entry1/instrument/manipulator/smdefocus",
-        "analyser_model": "FIXED_VALUE:Scienta DA30",
-        "analyser_slit_width": [
-            "entry1/instrument/analyser/entrance_slit_size",
-            "entry1/instrument/analyser_total/entrance_slit_size",
+        "analyser_model": lambda f: (
+            (
+                "FIXED_VALUE:MBS A1"
+                if parse(
+                    (
+                        (f["entry1/start_time"][()]).decode()
+                        if isinstance(f["entry1/start_time"][()], bytes)
+                        else f["entry1/start_time"][()]
+                    )
+                )
+                > datetime(2026, 4, 1, tzinfo=timezone.utc)
+                else "FIXED_VALUE:Scienta DA30"
+            )
+            if "entry1/start_time" in f
+            else None
+        ),
+        "analyser_slit_width": [  # If the slit size is 0, assume it's not set and return None, otherwise return the value.
+            lambda f: (
+                None
+                if (
+                    v := BaseHDF5DataLoader._extract_hdf5_value(
+                        f, "entry1/instrument/analyser/entrance_slit_size"
+                    )
+                )
+                == 0
+                else v
+            ),
+            lambda f: (
+                None
+                if (
+                    v := BaseHDF5DataLoader._extract_hdf5_value(
+                        f, "entry1/instrument/analyser_total/entrance_slit_size"
+                    )
+                )
+                == 0
+                else v
+            ),
         ],
         "analyser_slit_width_identifier": [
-            "entry1/instrument/analyser/entrance_slit_setting",
-            "entry1/instrument/analyser_total/entrance_slit_setting",
+            lambda f: (
+                None  # If the slit size is 0 which is physcially impossible, assume it's not set and return None
+                if BaseHDF5DataLoader._extract_hdf5_value(
+                    f, "entry1/instrument/analyser/entrance_slit_size"
+                )
+                == 0
+                else "entry1/instrument/analyser/entrance_slit_setting"  # otherwise return it
+            ),
+            lambda f: (
+                None
+                if BaseHDF5DataLoader._extract_hdf5_value(
+                    f, "entry1/instrument/analyser_total/entrance_slit_size"
+                )
+                == 0
+                else "entry1/instrument/analyser_total/entrance_slit_setting"
+            ),
         ],
         "analyser_eV": [
             "entry1/instrument/analyser/energies",
@@ -619,8 +664,38 @@ class I05NanoARPESLoader(I05ARPESLoader, BaseOpticsDataLoader):
[…]
