+++ CHANGELOG.md
@@ -5,7 +5,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [0.5.2]
+## [0.5.2.dev]
+
+### Changed
+
+- Diamond I05 data loader, following the analyser upgrade ([PR#XX](https://github.com/phrgab/peaks/pull/XX))
 
 ## [0.5.1] - 2026-05-29
 
