### 2026-07-07 243752a
fix plot_fit rendering in static docs builds
