### 2026-07-07 8cf2e7d
update i05 loader for nano upgrade; minor changes to HR metadata display
