+++ peaks/core/fileIO/loaders/diamond.py
@@ -193,8 +193,8 @@ class I05ARPESLoader(DiamondNXSLoader, BaseARPESDataLoader):
     }
 
     _analyser_name_conventions = {
-        "deflector_perp": "deflector_x",
-        "deflector_parallel": "deflector_y",
+        "deflector_perp": "deflector_x",  # These two don't exist for R4000 but fine for now
+        "deflector_parallel": "detector_y",  # Should handle them properly when we do the refactor
         "eV": ["energies", "kinetic_energy_center"],
         "theta_par": "angles",
         "hv": ["energy", "value"],
@@ -225,8 +225,36 @@ class I05ARPESLoader(DiamondNXSLoader, BaseARPESDataLoader):
             if "entry1/start_time" in f
             else None
         ),
-        "analyser_slit_width": "entry1/instrument/analyser/entrance_slit_size",
-        "analyser_slit_width_identifier": "entry1/instrument/analyser_total/entrance_slit_setting",
+        # "analyser_slit_width": "entry1/instrument/analyser/entrance_slit_size",
+        "analyser_slit_width": lambda f: (  # If the slit size is 0, assume it's not set and return None, otherwise return the value.
+            None
+            if (
+                v := BaseHDF5DataLoader._extract_hdf5_value(
+                    f, "entry1/instrument/analyser/entrance_slit_size"
+                )
+            )
+            == 0
+            else v
+        ),
+        # "analyser_slit_width_identifier": "entry1/instrument/analyser_total/entrance_slit_setting",
+        "analyser_slit_width_identifier": [
+            lambda f: (
+                None  # If the slit size is 0 which is physcially impossible, assume it's not set and return None
+                if BaseHDF5DataLoader._extract_hdf5_value(
+                    f, "entry1/instrument/analyser/entrance_slit_size"
+                )
+                == 0
+                else "entry1/instrument/analyser/entrance_slit_setting"  # otherwise return it
+            ),
+            lambda f: (
+                None
+                if BaseHDF5DataLoader._extract_hdf5_value(
+                    f, "entry1/instrument/analyser_total/entrance_slit_size"
+                )
+                == 0
+                else "entry1/instrument/analyser_total/entrance_slit_setting"
+            ),
+        ],
         "analyser_eV": "entry1/instrument/analyser/energies",
         "analyser_step_size": lambda f: (
             np.ptp(f["entry1/instrument/analyser/energies"])
@@ -244,8 +272,37 @@ class I05ARPESLoader(DiamondNXSLoader, BaseARPESDataLoader):
         "analyser_lens_mode": "entry1/instrument/analyser/lens_mode",
         "analyser_acquisition_mode": "entry1/instrument/analyser/acquisition_mode",
         "analyser_eV_type": "FIXED_VALUE:kinetic",
-        "analyser_deflector_parallel": "entry1/instrument/deflector_y/deflector_y",
-        "analyser_deflector_perp": "entry1/instrument/deflector_x/deflector_x",
+        "analyser_deflector_parallel": lambda f: (
+            (
+                "entry1/instrument/analyser/detector_y"
+                if parse(
+                    (
+                        (f["entry1/start_time"][()]).decode()
+                        if isinstance(f["entry1/start_time"][()], bytes)
+                        else f["entry1/start_time"][()]
+                    )
+                )
+                > datetime(2021, 7, 1, tzinfo=timezone.utc)
+                else None
+            )
+            if "entry1/start_time" in f
+            else None
+        ),
+        "analyser_deflector_perp": lambda f: (
+            (
[…]
