### 2026-07-07 13e8394
minor refactor to share start-time parsing across i05 nano loader
