### 2026-07-07 f8c2752
fix the aspect ratio and shorten the over-long title in the fit plots
