### 2026-07-07 855b788
wiring the time parsing logic into the parent class so im not repeating myself
