+++ peaks/core/fileIO/loaders/diamond.py
@@ -165,6 +165,28 @@ class DiamondNXSLoader(BaseHDF5DataLoader):
             f"and no signal attribute found to mark primary data in the group."
         )
 
+    @staticmethod
+    def _parse_start_time(f):
+        """Return the tz-aware acquisition start time."""
+        if "entry1/start_time" not in f:
+            return None
+        _raw_start_time = f["entry1/start_time"][()]
+        _raw_start_time = (
+            _raw_start_time.decode()
+            if isinstance(_raw_start_time, bytes)
+            else str(_raw_start_time)
+        )
+        start_time = parse(_raw_start_time)
+        if start_time.tzinfo is None:
+            start_time = start_time.replace(tzinfo=timezone.utc)
+        return start_time
+
+    @classmethod
+    def _get_start_time(cls, fpath):
+        """Parse the acquisition start time from the file at `fpath`."""
+        with h5py.File(fpath, "r") as f:
+            return cls._parse_start_time(f)
+
 
 @register_loader
 class I05ARPESLoader(DiamondNXSLoader, BaseARPESDataLoader):
@@ -210,20 +232,11 @@ class I05ARPESLoader(DiamondNXSLoader, BaseARPESDataLoader):
         "manipulator_x3": "entry1/instrument/manipulator/saz",
         "manipulator_salong": "entry1/instrument/manipulator/salong",
         "analyser_model": lambda f: (
-            (
-                "FIXED_VALUE:MBS A1"
-                if parse(
-                    (
-                        (f["entry1/start_time"][()]).decode()
-                        if isinstance(f["entry1/start_time"][()], bytes)
-                        else f["entry1/start_time"][()]
-                    )
-                )
-                > datetime(2021, 7, 1, tzinfo=timezone.utc)
-                else "FIXED_VALUE:Scienta R4000"
-            )
-            if "entry1/start_time" in f
-            else None
+            None
+            if (ac_date := DiamondNXSLoader._parse_start_time(f)) is None
+            else "FIXED_VALUE:MBS A1"
+            if ac_date > I05ARPESLoader._hr_analyser_upgrade_date
+            else "FIXED_VALUE:Scienta R4000"
         ),
         "analyser_slit_width": lambda f: (  # If the slit size is 0, assume it's not set and return None, otherwise return the value.
             None
@@ -271,19 +284,9 @@ class I05ARPESLoader(DiamondNXSLoader, BaseARPESDataLoader):
         "analyser_acquisition_mode": "entry1/instrument/analyser/acquisition_mode",
         "analyser_eV_type": "FIXED_VALUE:kinetic",
         "analyser_deflector_parallel": lambda f: (
-            (
-                "entry1/instrument/analyser/detector_y"
-                if parse(
-                    (
-                        (f["entry1/start_time"][()]).decode()
-                        if isinstance(f["entry1/start_time"][()], bytes)
-                        else f["entry1/start_time"][()]
-                    )
-                )
-                > datetime(2021, 7, 1, tzinfo=timezone.utc)
-                else None
-            )
-            if "entry1/start_time" in f
+            "entry1/instrument/analyser/detector_y"
+            if (ac_date := DiamondNXSLoader._parse_start_time(f))
+            and ac_date > I05ARPESLoader._hr_analyser_upgrade_date
             else None
         ),
         "analyser_deflector_perp": lambda f: (
@@ -292,13 +295,8 @@ class I05ARPESLoader(DiamondNXSLoader, BaseARPESDataLoader):
                 if "entry1/instrument/deflector_x/deflector_x" in f
[…]
