### 2026-07-07 12809ca
bump main version to 0.5.2
