+++ CHANGELOG.md
@@ -5,6 +5,16 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.5.3.dev]
+
+### Added
+
+### Fixed
+
+### Changed
+
+### Removed
+
 ## [0.5.2] - 2026-07-07
 
 ### Added
