### 2026-07-08 b01a3a1
change tqdm.notebook to tpqm.auto to make progress bars work outside jupyter; warnings support marimo
