### 2026-07-08 96b4e59
add unit tests for marimo callouts and update CL
