### 2026-07-08 334bcfc
preserve aspect ratio for wide figures from plot_grid in marimo
