### 2026-07-14 e122086
fix rounding being 0 when building colourbar
