+++ peaks/core/GUI/disp_panels/disp_2d.py
@@ -414,7 +414,6 @@ class _Disp2D(QtWidgets.QMainWindow):
             colorMapMenu=True,
             limits=(self.c_min, self.c_max),
             rounding=min(c_span / 2000, 1) if c_span > 0 else 1e-9,
-            # rounding=min(abs(self.c_max - self.c_min) / 2000, 1),
             interactive=True,
             orientation="h",
             values=c_range,
