+++ CHANGELOG.md
@@ -14,8 +14,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Fixed
 
 - `plot_grid` renders with a distorted aspect ratio in marimo notebooks when the figure is wider than the notebook column ([PR#56](https://github.com/phrgab/peaks/pull/56))
-- `ZeroDivisionError` in `disp` colour bar for constant-value data slices
-- Metadata reports `[nan, nan]` when a scan contains a few NaN readbacks
+- `ZeroDivisionError` in `disp` colour bar for constant-value data slices ([PR#57](https://github.com/phrgab/peaks/pull/57))
+- Metadata reports `[nan, nan]` when a scan contains a few NaN readbacks ([PR#57](https://github.com/phrgab/peaks/pull/57))
 
 ### Changed
 
