### 2026-07-15 1a43949
remove hard-coded defocus dim in plot_nanofocus
