### 2026-07-15 ce6ca17
guess focus dim if absent
