### 2026-07-15 3930b3d
trim repeated data points at the end of travel in i05 ana maps
