### 2026-07-15 d0d93fb
handle clipboard failiures gracefully in disp panels
