### 2026-07-18 39d1dea
error handling when a multi-leaf dt doesnt contain 2D data
