### 2026-07-20 e942510
Update CHANGELOG.md
Co-authored-by: Phil King <78075132+phrgab@users.noreply.github.com>
