### 2026-07-24 3fba611
fix lazy-loading condition in netcdf loader and enable chunking therr
