### 2026-07-26 287e64e
Centralise Qt viewer lifecycle management
Move shared Qt setup into a single runtime module used by the 2D, 3D,
and 4D display panels. Automatically enable Qt6 event-loop integration
in IPython and Jupyter when no conflicting GUI loop is active, while
retaining blocking QApplication.exec() behaviour for regular Python
scripts.

Store active viewers on the shared QApplication instance so windows
remain alive without callers retaining references, and remove them from
the registry when destroyed. Remove duplicated application creation,
IPython detection, event-loop handling, and close-event logic
[…]
