### 2026-07-27 85073c4
pin mathjax version to 3 in sphinx config for now due to bokeh imcompatibility
