### 2026-07-27 fd552b2
limiting number of open viewers by a pks option
