+++ peaks/core/options.py
@@ -169,6 +169,88 @@ class FileIOOptions:
         return {k.lstrip("_"): raw_dict[k] for k in raw_dict}
 
 
+_DEFAULT_MAX_VIEWERS = 1
+
+
+class GuiOptions:
+    """Options controlling interactive display panel GUIs.
+
+    This class allows setting the maximum number of display panels or other GUIs
+    that may be open simultaneously. Setting ``max_viewers`` to ``None`` disables
+    the limit. The reset method restores the default display options.
+
+    Examples
+    --------
+    Example usage is as follows::
+
+        import peaks as pks
+
+        # Set the maximum number of simultaneously open display panels
+        pks.opts.gui.max_viewers = 5
+
+        # Disable the display panel limit
+        pks.opts.gui.max_viewers = None
+
+        # Show display options
+        pks.opts.gui
+
+        # Reset all Gui options
+        pks.opts.gui.reset()
+
+        # Reset all options
+        pks.opts.reset()
+    """
+
+    def __init__(self):
+        self._max_viewers = _DEFAULT_MAX_VIEWERS
+
+    @property
+    def max_viewers(self):
+        """Return the maximum number of simultaneously open display panels."""
+        return self._max_viewers
+
+    @max_viewers.setter
+    def max_viewers(self, value):
+        if value is None:
+            self._max_viewers = None
+        elif isinstance(value, bool):
+            raise TypeError("Maximum viewers must be a positive integer or None.")
+        elif isinstance(value, int):
+            if value < 1:
+                raise ValueError("Maximum viewers must be greater than or equal to 1.")
+            self._max_viewers = value
+        else:
+            raise TypeError("Maximum viewers must be a positive integer or None.")
+
+    @max_viewers.deleter
+    def max_viewers(self):
+        self._max_viewers = _DEFAULT_MAX_VIEWERS
+
+    def reset(self):
+        """Reset display options to their defaults."""
+        self._max_viewers = _DEFAULT_MAX_VIEWERS
+
+    def __repr__(self):
+        """Return a string representation of the display options."""
+        return format_colored_dict(self.dict())
+
+    def set(self, **kwargs):
+        """Set display options."""
+        if "max_viewers" in kwargs:
+            self.max_viewers = kwargs.pop("max_viewers")
+
+        if kwargs:
+            raise ValueError(
+                f"Invalid keyword argument(s): {set(kwargs.keys())}. "
+                f"Expected options from {set(self.dict().keys())}"
+            )
+
+    def dict(self):
+        """Return a dictionary representation of the display options."""
+        raw_dict = vars(self)
+        return {k.lstrip("_"): raw_dict[k] for k in raw_dict}
+
+
 class Options:
     """
     Singleton class to hold all fixed option groups, such as FileIO.
@@ -180,6 +262,9 @@ class Options:
     ----------
     FileIO : FileIOOptions
         An instance of the FileIOOptions class to manage file input/output settings.
+
+    gui : GuiOptions
+        Options controlling interactive display panels.
 
     Methods
     -------
@@ -197,6 +282,16 @@ class Options:
         pks.opts.FileIO.ext = ['nxs', 'zip']  # Default extensions
         pks.opts.FileIO.loc = 'Diamond_I05_nano-ARPES'  # loc to use
         pks.opts.FileIO.lazy_size = 500000000  # Set lazy size to 500 Mb
+
+        # Allow a fixed number of display panels to be open simultaneously
+        pks.opts.gui.max_viewers = 3
+
+        # Disable the display-panel limit
+        pks.opts.gui.max_viewers = None
+
+        # Reset display options
+        pks.opts.gui.reset()  # Disables multiple panels by defualt
+
 
         # Display all the current options
         pks.opts
@@ -217,6 +312,8 @@ class Options:
[…]
