+++ CHANGELOG.md
@@ -28,6 +28,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Changed
 
 - Limit font size and max lines of titles in `plot_fit` outputs ([PR#56](https://github.com/phrgab/peaks/pull/56))
+- Automatically enable Qt6 event-loop integration when running from Jupyter/IPython, with simultaneous viewer management handled by a `pks.opt` ([PR#68](https://github.com/phrgab/peaks/pull/68))
 
 ### Removed
 
