### 2026-07-28 dccacbf
add support for i05 fly spatial maps
