+++ peaks/core/fileIO/loaders/diamond.py
@@ -187,6 +187,37 @@ class DiamondNXSLoader(BaseHDF5DataLoader):
         with h5py.File(fpath, "r") as f:
             return cls._parse_start_time(f)
 
+    @staticmethod
+    def _trim_sweep_artefacts(da, dim):
+        """Trim points at either end of a fly-scan axis where the motor was still
+        accelerating or decelerating, leaving non-monotonic/repeated coordinate values."""
+        val = np.asarray(da[dim].values, dtype=float)
+        if val.size < 3:
+            return da
+        steps = np.diff(val)
+        median_step = np.median(steps)
+        if not np.isfinite(median_step) or median_step == 0:
+            return da
+        threshold = 0.1 * np.abs(median_step)  # for motor noise
+        direction = np.sign(median_step)
+        moving = np.flatnonzero(
+            (np.abs(steps) > threshold) & (np.sign(steps) == direction)
+        )
+        if not moving.size:
+            return da
+        first_good = int(moving[0])
+        last_good = int(moving[-1]) + 1
+        n_trim = val.size - (last_good - first_good + 1)
+        if n_trim:
+            da = da.isel({dim: slice(first_good, last_good + 1)})
+            analysis_warning(
+                f"Non-monotonic values detected in `{dim}` at the start and/or end of "
+                f"travel. Trimmed {n_trim} data point(s).",
+                "warning",
+                "Removed non-monotonic data",
+            )
+        return da
+
 
 @register_loader
 class I05ARPESLoader(DiamondNXSLoader, BaseARPESDataLoader):
@@ -216,7 +247,7 @@ class I05ARPESLoader(DiamondNXSLoader, BaseARPESDataLoader):
 
     _analyser_name_conventions = {
         "deflector_perp": "deflector_x",  # These two don't exist for R4000 but fine for now
-        "deflector_parallel": "detector_y",  # Should handle them properly when we do the refactor
+        "deflector_parallel": "deflector_y",  # Should handle them properly when we do the refactor. This was called detector_y at least before July 2026
         "eV": ["energies", "kinetic_energy_center"],
         "theta_par": "angles",
         "hv": ["energy", "value"],
@@ -228,7 +259,10 @@ class I05ARPESLoader(DiamondNXSLoader, BaseARPESDataLoader):
         "manipulator_tilt": "entry1/instrument/manipulator/satilt",
         "manipulator_azi": "entry1/instrument/manipulator/saazimuth",
         "manipulator_x1": "entry1/instrument/manipulator/sax",
-        "manipulator_x2": "entry1/instrument/manipulator/saz",
+        "manipulator_x2": [
+            "entry1/instrument/saz_fly/value",  # fly-scan path must come first
+            "entry1/instrument/manipulator/saz",
+        ],
         "manipulator_x3": "entry1/instrument/manipulator/say",
         "manipulator_salong": "entry1/instrument/manipulator/salong",
         "analyser_model": lambda f: (
@@ -284,7 +318,9 @@ class I05ARPESLoader(DiamondNXSLoader, BaseARPESDataLoader):
         "analyser_acquisition_mode": "entry1/instrument/analyser/acquisition_mode",
         "analyser_eV_type": "FIXED_VALUE:kinetic",
         "analyser_deflector_parallel": lambda f: (
-            "entry1/instrument/analyser/detector_y"
+            "entry1/instrument/analyser/deflector_y"
+            if "entry1/instrument/analyser/deflector_y" in f
+            else "entry1/instrument/analyser/detector_y"
             if (ac_date := DiamondNXSLoader._parse_start_time(f))
             and ac_date > I05ARPESLoader._hr_analyser_upgrade_date
             else None
[…]
