### 2026-07-28 94ee949
add missing units to i05 deflector metadata
