### 2026-07-29 cf6a2fe
add comments noting 2D transmission scan units are in millimetres
