### 2026-07-29 a3f3541
add debug prints to test CI
