### 2026-07-29 3c753ad
fire a warning if local mirror path in invalid
