### 2026-07-29 a2f7cc7
check whether ci env var is set to true during CI runs
