+++ peaks/core/utils/sample_data.py
@@ -35,15 +35,18 @@ class ZenodoDownloader:
         self.file_list = file_list
         self.token = token or os.getenv("ZENODO_TOKEN")
         self.local_mirror = os.getenv("LOCAL_MIRROR_PATH")
-        if self.local_mirror and not os.path.isdir(self.local_mirror):
-            analysis_warning(
-                f"LOCAL_MIRROR_PATH is set to <code>{self.local_mirror}</code>, but this directory does not exist. "
-                "Files will be downloaded from Zenodo instead.",
-                title="Invalid LOCAL_MIRROR_PATH",
-                warn_type="warning",
-            )
+        self._in_ci = bool(os.getenv("CI"))
         self._tempdir_context = None
         self.downloaded_files = {}
+
+    # def _ci_log(self, message):
+    def _ci_log(self):
+        """Log plain messages in CI environment."""
+        # if self._in_ci:
+        print(
+            f"[DEBUG] CI env var is {os.getenv('CI')!r}, self._in_ci = {self._in_ci}",
+            flush=True,
+        )
 
     def _get_local_path(self, filename):
         """Check if local mirror available and if files exist there."""
