### 2026-07-29 7250825
add debug call
