### 2026-07-29 aa503bd
log local mirror hit/miss and warn on invalid path
