### 2026-07-29 fd3aafc
add more COD mirrors
