### 2026-07-29 154a4b4
add ci logging in cif file download
