### 2026-07-29 e0bff58
Apply suggestions from code review
Co-authored-by: Shu <148200713+shum23@users.noreply.github.com>
