### 2026-08-03 0067d82
fix iplot sliders being static/broken in docs build
