### 2026-08-03 b3bf31d
add a citation
