+++ paper/paper.bib
@@ -185,6 +185,18 @@
 	journal = {Journal of Open Source Software},
 }
 
+@article{Pudelko2025,
+    author = {Pudelko, Wojciech R. and Kramer, Kevin and Küspert, Julia and Chang, Johan and Plumb, Nicholas C.},
+    doi = {10.21105/joss.09129},
+    url = {https://doi.org/10.21105/joss.09129},
+    journal = {Journal of Open Source Software},
+    number = {114},
+    pages = {9129},
+    title = {{PIVA: Photoemission Interface for Visualization and Analysis}},
+    volume = {10},
+    year = {2025}
+}
+
 @misc{ErLabPy,
 	author = {Higashihira Han, Kimoon},
 	title = {erlab: A Complete Python Workflow for ARPES Experiments},
