### 2026-08-03 331ff52
raise import error when trimesh is missing
