### 2026-08-03 5516b65
hardcode the cif file and use it when present
