+++ docs/source/files/cif_file_credits.md
@@ -1,0 +1,7 @@
+Structure of PdCoO2, from:
+Guowei Li, Seunghyun Khim, Celesta S. Chang, Chenguang Fu, Nabhanila Nandi, Fan Li, Qun Yang, Graeme R. Blake, Stuart Parkin, Gudrun Auffermann, Yan Sun, David A. Muller, Andrew P. Mackenzie, Claudia Felser;
+"In Situ Modification of a Delafossite-Type PdCoO2 Bulk Single Crystal for Reversible Hydrogen Sorption and Fast Hydrogen Evolution",
+ACS Energy Lett. 13 September 2019; 4 (9): 2185–2191.
+DOI: 10.1021/acsenergylett.9b01527
+URL: https://doi.org/10.1021/acsenergylett.9b01527
+Licensed CC BY 4.0. Retrieved via the Crystallography Open Database (COD entry 4515175).
