### 2026-08-04 9977eb9
wording tweaks
