### 2026-08-04 6ce986a
Update CHANGELOG.md
Co-authored-by: Phil King <78075132+phrgab@users.noreply.github.com>
