+++ CHANGELOG.md
@@ -30,7 +30,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Cap embedded slider states in docs builds ([PR#67](https://github.com/phrgab/peaks/pull/67))
 - Missing units on the deflector metadata for I05 data that can break k-conversion ([PR#71](https://github.com/phrgab/peaks/pull/71))
 - Fix sample-data source logging leaking into built docs - gate it behind a separate opt-in env var ([PR#73](https://github.com/phrgab/peaks/pull/73))
-- Raises a clear error if `trimesh` is missing when importing the structure module ([PR#76](https://github.com/phrgab/peaks/pull/76))
+- Raises a clear error if `ase` or `trimesh` is missing when importing the structure module ([PR#76](https://github.com/phrgab/peaks/pull/76))
 - Static inline interactive `iplot` in docs. Multi-dimensional `iplot`s now each have a few frames to animate over ([PR#76](https://github.com/phrgab/peaks/pull/76))
 
 ### Changed
