### 2026-08-24 7583bc3
a bit more clarification on k-conv in the docs
