### 2026-08-31 8fd3424
the initial version provided by Jacek Osiecki at Bloch, Max IV
