+++ peaks/core/fileIO/loaders/bloch.py
@@ -1,9 +1,27 @@
 """Functions to load data from the Bloch beamline at MAX IV Laboratory."""
 
+from __future__ import annotations
+
+from pathlib import Path
+from typing import Any
+
+import h5py
+import numpy as np
 import pint_xarray
 
-from peaks.core.fileIO.base_arpes_data_classes.base_ses_class import SESDataLoader
+from peaks.core.fileIO.base_arpes_data_classes.base_arpes_data_class import (
+    BaseARPESDataLoader,
+    ureg,
+)
+from peaks.core.fileIO.base_data_classes.base_hdf5_class import BaseHDF5DataLoader
 from peaks.core.fileIO.loc_registry import register_loader
+
+try:  # Optional; only needed when pks.load(..., lazy=True)
+    import dask.array as dask_array
+except Exception:  # pragma: no cover
+    dask_array = None
+
+from peaks.core.fileIO.base_arpes_data_classes.base_ses_class import SESDataLoader
 
 ureg = pint_xarray.unit_registry
 
@@ -36,3 +54,1554 @@ class BlochArpesLoader(SESDataLoader):
         f"manipulator_{dim}": ("mm" if dim in ["x1", "x2", "x3"] else "deg")
         for dim in _manipulator_name_conventions.keys()
     }
+
+
+_STRING_TYPES = (bytes, np.bytes_)
+
+
+def _decode(value: Any) -> Any:
+    """Decode HDF5 bytes/numpy scalars into ordinary Python values."""
+    if isinstance(value, _STRING_TYPES):
+        return value.decode(errors="replace")
+    if isinstance(value, np.ndarray):
+        if value.shape == ():
+            return _decode(value.item())
+        if value.size == 1:
+            return _decode(value.reshape(-1)[0])
+        return [_decode(v) for v in value.tolist()]
+    if hasattr(value, "item") and not isinstance(value, (str, bytes)):
+        try:
+            return value.item()
+        except Exception:
+            return value
+    return value
+
+
+def _as_text(value: Any, default: str = "") -> str:
+    value = _decode(value)
+    if value is None:
+        return default
+    return str(value)
+
+
+def _read_scalar(group_or_file: Any, path: str, default: Any = None) -> Any:
+    """Read a scalar dataset if it exists; otherwise return default."""
+    try:
+        node = group_or_file[path]
+    except Exception:
+        return default
+    try:
+        return _decode(node[()])
+    except Exception:
+        return default
+
+
+def _read_first_scalar(group_or_file: Any, paths: list[str], default: Any = None) -> Any:
+    """Read the first non-empty scalar from a list of HDF5 paths."""
+    for path in paths:
+        value = _read_scalar(group_or_file, path, None)
+        if value not in (None, ""):
+            return value
+    return default
+
+
+def _read_numeric(group_or_file: Any, paths: list[str], default: Any = None) -> Any:
+    """Read the first numeric scalar from a list of HDF5 paths."""
+    value = _read_first_scalar(group_or_file, paths, default=None)
+    if value in (None, ""):
+        return default
+    try:
+        return float(value)
+    except Exception:
+        return default
+
+
+def _read_vector(dataset: Any, fallback_length: int) -> np.ndarray:
+    if dataset is None:
+        return np.arange(int(fallback_length), dtype=np.float64)
+    try:
+        values = np.asarray(dataset[()], dtype=np.float64).reshape(-1)
+        if values.size == int(fallback_length):
+            return values
+    except Exception:
+        pass
+    return np.arange(int(fallback_length), dtype=np.float64)
+
+
[…]
