### 2026-08-31 e0a3f48
working-ish version before the beamtime for loading dispersions, defl maps, fly spatial maps nad hv scans
