### 2026-09-10 6e1f596
JOSS manuscript tweaks based on editor's comments
