+++ paper/paper.bib
@@ -60,7 +60,7 @@
 	url = {https://www.nature.com/articles/s41535-021-00407-5},
 	doi = {10.1038/s41535-021-00407-5},
 	urldate = {2025-07-15},
-	journal = {npj Quantum Mater.},
+	journal = {{npj} Quantum Mater.},
 	author = {Iwasawa, Hideaki and Ueno, Tetsuro and Masui, Takahiko and Tajima,
 	          Setsuko},
 	year = {2022},
@@ -165,7 +165,7 @@
 
 @misc{das,
 	author = {P. Das},
-	title = {ARPES Python Tools},
+	title = {ARPES {P}ython {T}ools},
 	year = {2026},
 	publisher = {GitHub},
 	journal = {GitHub repository},
@@ -182,14 +182,14 @@
 	pages = {2969},
 	author = {Kramer, Kevin and Chang, Johan},
 	title = {Visualization of Multi-Dimensional Data -- The data-slicer Package},
-	journal = {Journal of Open Source Software},
+	journal = {J. Open Source Softw.},
 }
 
 @article{Pudelko2025,
     author = {Pudelko, Wojciech R. and Kramer, Kevin and Küspert, Julia and Chang, Johan and Plumb, Nicholas C.},
     doi = {10.21105/joss.09129},
     url = {https://doi.org/10.21105/joss.09129},
-    journal = {Journal of Open Source Software},
+    journal = {J. Open Source Softw.},
     number = {114},
     pages = {9129},
     title = {{PIVA: Photoemission Interface for Visualization and Analysis}},
@@ -199,7 +199,7 @@
 
 @misc{ErLabPy,
 	author = {Higashihira Han, Kimoon},
-	title = {erlab: A Complete Python Workflow for ARPES Experiments},
+	title = {{erlab}: {A} {Complete} {P}ython {Workflow} for {ARPES} {Experiments}},
 	year = {2026},
 	publisher = {Zenodo},
 	url = {https://doi.org/10.5281/zenodo.18392129},
@@ -207,7 +207,7 @@
 }
 
 @article{hoyer2017xarray,
-	title = {xarray: {N-D} labeled arrays and datasets in {Python}},
+	title = {{xarray}: {N-D} labeled arrays and datasets in {Python}},
 	author = {Hoyer, S. and J. Hamman},
 	journal = {Journal of Open Research Software},
 	volume = {5},
@@ -232,8 +232,8 @@
 	          Hewitt, David and Hall, Alex and Plot, Victorien},
 	title = {{Pydantic Validation}},
 	url = {https://github.com/pydantic/pydantic},
-	version = {v2.13.0a0+dev},
-	year = {2025},
+	version = {v2.13.4},
+	year = {2026},
 }
 
 @misc{pint,
@@ -252,7 +252,7 @@
 	title = {{LMFIT: Non-Linear Least-Squares Minimization and Curve-Fitting for
 	         Python}},
 	version = {1.3.3},
-	url = {https://github.com/lmfit/lmfit-py/},
+	doi = {10.5281/zenodo.15014437},
 	year = {2025},
 }
 
@@ -262,7 +262,7 @@
 	          Thirion, B. and Grisel, O. and Blondel, M. and Prettenhofer, P. and
 	          Weiss, R. and Dubourg, V. and Vanderplas, J. and Passos, A. and
 	          Cournapeau, D. and Brucher, M. and Perrot, M. and Duchesnay, E.},
-	journal = {Journal of Machine Learning Research},
+	journal = {J. Mach. Learn. Res.},
 	volume = {12},
 	pages = {2825--2830},
[…]
