### 2026-09-10 1de7ac1
Update paper/paper.md
Co-authored-by: Phil King <78075132+phrgab@users.noreply.github.com>
