+++ paper/paper.md
@@ -39,7 +39,8 @@ bibliography: paper.bib
 
 # Summary
 
-The electronic band structure, describing the motion and interactions of electrons in materials, dictates the electrical, optical, and thermodynamic properties of solids. Angle-resolved photoemission spectroscopy (ARPES) provides a direct experimental probe of such electronic band structures, and so is widely employed in the study of functional, quantum, and 2D materials [@Damascelli2003; @king_angle_2021; @sobota_angle-resolved_2021]. `peaks` (**P**ython **E**lectron spectroscopy **A**nalysis by **K**ing group @ **S**t Andrews) provides a Python package for advanced data analysis of ARPES and related spectroscopic data. It facilitates the fast visualisation and analysis of multi-dimensional datasets, allows for the complex data hierarchy typical to ARPES experiments, and supports lazy data loading and parallel processing, reflecting the ever-increasing data volumes used in ARPES. It is designed to be run in an interactive notebook environment, with extensive inline and pop-out GUI support for data visualisation.
+
+The electronic structure of materials dictates their electrical, optical, and thermodynamic properties. Angle-resolved photoemission spectroscopy (ARPES) provides a direct, momentum-resolved, experimental probe of such electronic structures, and so is widely employed in the study of functional, quantum, and 2D materials [@Damascelli2003; @king_angle_2021; @sobota_angle-resolved_2021]. `peaks` (**P**ython **E**lectron spectroscopy **A**nalysis by **K**ing group @ **S**t Andrews) provides a Python package for advanced data analysis of ARPES and related spectroscopic data. It facilitates the fast visualisation and analysis of multi-dimensional datasets, allows for the complex data hierarchy typical to ARPES experiments, and supports lazy data loading and parallel processing, reflecting the ever-increasing data volumes used in ARPES. It is designed to be run in an interactive notebook environment, with extensive inline and pop-out GUI support for data visualisation.
 
 # Statement of need
 
