+++ paper/paper.md
@@ -39,7 +39,6 @@ bibliography: paper.bib
 
 # Summary
 
-
 The electronic structure of materials dictates their electrical, optical, and thermodynamic properties. Angle-resolved photoemission spectroscopy (ARPES) provides a direct, momentum-resolved, experimental probe of such electronic structures, and so is widely employed in the study of functional, quantum, and 2D materials [@Damascelli2003; @king_angle_2021; @sobota_angle-resolved_2021]. `peaks` (**P**ython **E**lectron spectroscopy **A**nalysis by **K**ing group @ **S**t Andrews) provides a Python package for advanced data analysis of ARPES and related spectroscopic data. It facilitates the fast visualisation and analysis of multi-dimensional datasets, allows for the complex data hierarchy typical to ARPES experiments, and supports lazy data loading and parallel processing, reflecting the ever-increasing data volumes used in ARPES. It is designed to be run in an interactive notebook environment, with extensive inline and pop-out GUI support for data visualisation.
 
 # Statement of need
@@ -72,7 +71,7 @@ Being primarily intended for use by experimental physicists, a high-level progra
 
 # Research impact statement
 
-The development of `peaks` has so far underpinned data analysis in more than 10 academic publications from the developers’ group (e.g. [@buchberger_persistence_2025; @armitage_electronic_2025; @mo_resonant_2026; @saika_flat_2025; @siemann_dichotomy_2025; @edwards_giant_2023]), during which time it has undergone continuous and active development and maintenance. Since its public release in 2025, we have been contacted by several groups who are starting to use `peaks` as their primary analysis driver for ARPES, reflecting increased community adoption (e.g. [@peaks_pr3]), and so we expect its impact to substantially broaden over the coming years.
+The development of `peaks` has so far underpinned data analysis in more than 10 academic publications from the developers’ group [e.g. @armitage_electronic_2025; @buchberger_persistence_2025; @mo_resonant_2026; @saika_flat_2025; @siemann_dichotomy_2025; @edwards_giant_2023], during which time it has undergone continuous and active development and maintenance. Since its public release in 2025, we have been contacted by several groups who are starting to use `peaks` as their primary analysis driver for ARPES, reflecting increased community adoption [e.g. @peaks_pr3], and so we expect its impact to substantially broaden over the coming years.
 
 To support broader adoption, `peaks` provides detailed documentation covering installation instructions and a thorough user guide which is based on rendered Jupyter notebook outputs. These are written in a didactic style for users who may be familiar with ARPES data analysis but not Python, while a single-line installation from PyPI, including all required dependencies, makes installation simple for new users. The Jupyter notebooks which make up the documentation user guide are available in the `peaks` GitHub repository, and can be run locally, with automated download of sample data from an open-source Zenodo repository to facilitate training. These notebooks also serve as a test set for `peaks`, which are validated in CI for all new contributions made to `peaks` before they are merged into the main repository branch. The package uses semantic versioning, with tagged commits for each version and corresponding releases on PyPI. Software API and contribution guidelines are included in the documentation, and a standard code style is enforced, facilitating broader contributions - which we strongly welcome.
 
