### 2026-09-15 2e3f85d
fix moving crosshair with arrow keys doesnot update dispersion in spatial disp
