### 2026-09-15 c982cd2
fix ibw axis scaling that affects spatial map readouts
