### 2026-09-15 25a4eb0
scale the readout precision ofr primary dims in the 4D disp to the ax step size
