### 2026-09-15 7edb762
linting - deleted unused variable
