+++ CHANGELOG.md
@@ -9,6 +9,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
+- (Alpha) Support for loading NeXus data from the Bloch beamline at MAX IV. Tested on data acquired in August 2026, covering fly-scan spatial maps, photon-energy scans, deflector maps and fixed/swept dispersions ([PR#85](https://github.com/phrgab/peaks/pull/85))
+
 ### Fixed
 
 ### Changed
